[roll] Roll fuchsia [ffx][log] Allow filtering on the moniker last segment with --tag
When a component has no tags, the user may see the last segment of the moniker
as the tag. In this cases, it's confusing that `--tag foo` doesn't work
when the user sees `[foo]` in the logs, but `--filter foo` and
`--moniker core/foo` work.

This change allows applying --tag to the last segment of monikers.

Tested: unit test + local verification with power-broker
Original-Fixed: 336879224
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1039932
Original-Revision: e0bc3d4e34fd148af1920621a3bb84ad2ee45fc3
GitOrigin-RevId: b8be4bda061d278c1aad1c0a1665ccb28381f952
Change-Id: I3abcf3b865fca11607ed85ba5c1083396412ac60