[roll] Roll fuchsia [aml][pwm] Apply channel config during init
This is a bug where the channel config passed in from the board driver
is not applied during init and therefore incorrect config was passed on
to the pwm clients as well. This resulted in pwm period being set to 0
for both voltage regulators in vim3.

Also added logs in aml-pwm-regulator driver to fail early when invalid
period is passed.

Original-Bug: 338041521
Test: Manually checked the config values.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1039915
Original-Revision: 18e86439bf988298e52fd26aca60f7a3149e6b8f
GitOrigin-RevId: 7adbb0ea078c1701e4419a2d5e2bd0f9dd1235fd
Change-Id: Id0b3daee25df1e6c82d0382c97025bb83eba9317