[roll] Roll fuchsia [devshell][serve] Fix setting the port
The change in fxr/1038492 changed which positional arg the arg parsing
was reading to find the port in a way that resulted in it usually being
unset. This restores the functionality by making the behaviour more like
the other arguments that take a second arg.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1040172
Original-Revision: 50a9457b7c6f50fe54f99fcb181066ecd34beb43
GitOrigin-RevId: 52f0dbbbf71cd890eeb31705b3e9d03d576602a1
Change-Id: I56e6d2c484250d54b6d486c0b26a7fe1c3256ce3