[roll] Roll fuchsia [components][dictionaries] Extension is on demand
Currently, when a dictionary that has `extends` is routed, dictionary
extension is performed once, and that extended dictionary is cached and
reused on subsequent routing requests.

Instead, this change makes it so that the extension is performed anew on
every route.

The reason for this change is to prepare for `extends program`, which
needs these semantics because the dictionary being extended will change
if the program is restarted. With this, `extends <other>` will be
consistent with `extends program`. It's also simpler to implement; the
only downside is the minor performance cost of creating a new dictionary
and routing the source dictionary on every routing request.

Original-Bug: 336363726
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1036512
Original-Revision: 29bfd17edb2d32bc38a4b968f66aed191d6f0f21
GitOrigin-RevId: 2e8b734391dc55795484d2b60671cf976c56395d
Change-Id: I9590d7453ead6649a54b73089e127f2063def069