[roll] Roll fuchsia [fake-display-stack] Replace CoordinatorDriver with non-driver Controller.
Previously in the fake-display-stack, the CoordinatorDriver was bound to
a mock device node. This change replaces the CoordinatorDriver with a
non-driver Controller instance; the binding between Controller and the
fake-dispaly engine is manually created.

This is part of the changes to migrate fake-display-stack out of the
driver framework v1.

BUg: 338140042
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1040474
Original-Revision: a589d84d389b8146711fccac35717eb53703923c
GitOrigin-RevId: 1d4df0a4c05c4909133aea914d48ca1dbb976760
Change-Id: I9c5ac64be387c83f4d261f63f2cd42baa7a1393b