[roll] Roll fuchsia [cml] Replace merge macro with generic function
Replace the macro with a generic function. Instead of
calling this on every capability type, we can simply check that
the types match and that the names are the same.

This gives us better editor support for this function and
improve readability.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1040232
Original-Revision: e4c3cb44d692c6a6c3989f21509214c51e9662b0
GitOrigin-RevId: 2a13662d1ee67cb13d542aed6f548897646fc568
Change-Id: I586f07f9aa8cc76446d7b5f4785799f29d9b4ce9