[roll] Roll third_party/shac-project/checks-starlark [shac] Make default buildifier tool parameter a function call.
In fuchsia.git, shac checks need access to the 'ctx' variable
in order to determine the location of the prebuilt tools.

This changes the default tool parameter from a string to
a lambda, so this check can be used in both environments.

Tested by pinning this change in fuchsia.git and running the tools.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/shac-project/checks-starlark/+/1040133
Original-Revision: d1ba7bb7888c83ceb923c7e9ebafb8d84ce2ef16
GitOrigin-RevId: c4f7be236c3a7c3d278ea1ec0dbbb009ee492a76
Change-Id: I3030e342fdb7634a99b309653a9733a07ee5c665