[roll] Roll third_party/shac-project/checks-cpp [shac] Make default clang-format tool parameter a function call.
In fuchsia.git, shac checks need access to the 'ctx' variable
in order to determine the location of the prebuilt tools.

This changes the default tool parameter from a string to
a lambda, so this check can be used in both environments.

Tested by pinning this change in fuchsia.git and running the tools.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/shac-project/checks-cpp/+/1040134
Original-Revision: 85b972328c5868effcd584758b9ca4bf7f800f8f
GitOrigin-RevId: fadb14878aeab9d2bce0f126b0c5ecea1d72794d
Change-Id: I19919123f68d38eff710ef4ef86ed43c06789dbd