[roll] Roll fuchsia [netstack3] Prepare for deferred address removal
We can't assume that the primary reference for addresses can always be
dropped with no strong references held anywhere. Prepare the world for
that by allowing addresses to enter a "pre-removal" state where the
config has been extracted out.

This is somewhat an artifact of how SLAAC works - it wants to be
notified *after* the address was removed and it wants to own the
configuration.

Original-Bug: 336291808
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1040180
Original-Revision: f33fc2893db3fe91776e4640aea3c29475625d50
GitOrigin-RevId: eed8201d83f7c45352d7660829b595d4cc36130b
Change-Id: I71b7fa7e25b18f383f0f76ef8bda65a6d7f9d128