[roll] Roll fuchsia [kernel][arm][mmu] skip break-before-make during early kernel boot
This allows the kernel to split its own physmap and its own kernel
mappings' large pages while the system is still running on the bootup
core.

Break-before-make causes some issues during early kernel boot,
specifically when the kernel unmaps sections of the physmap and
re-protects sections of the kernel image itself. In these cases based on
how the kernel address space was set up initially it may need to split
pages. In various scenarios this could result in either temporarily
unmapping the kernel itself while its running or unmapping a section of
the physmap that the mmu code needs to access the page tables.

For now, continuing to allow break-before-make gets the kernel over this
hump, though it's technically pushing the architecture a bit, since not
BBM when splitting a large page *may* cause a TLB conflict abort,
depending on the core implementation. There's no reason to believe any
cores that exist exhibit this behavior, but it may be a thing in the
future.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1039220
Original-Revision: ceecbb1490c3c819f218c1b93754a94cba38990b
GitOrigin-RevId: 7b32f9a07aaf1c3e22a89a12a25fa614fd873458
Change-Id: I8dadb641be9275e8ff1282451e0f0e6ede0266cc