[roll] Roll fuchsia [shac] Use shared shac checks in fuchsia.git.
This modifies the starlark.star file to use the buildifier format
check from the shared repository. There should be no change in
behavior.

This also registers the buildifier_lint check from the shared repo
configured as a warning. There are 32 files with linting issues
across the tree that I will address before enabling this
as blocking. (or if that is undesirable, I will try again to land a shac.star file in the //build/bazel_sdk directory)

This also also makes it possible to run the clang-format check from
the shared repository, though it results in an error (b/b/338231347)

Original-Bug: b/338231347
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1040164
Original-Revision: 7000b97e25b8a808771f2b086c54ede3852311e8
GitOrigin-RevId: b733406cce65d68101f52ead77505f790a86d422
Change-Id: Ia28b8957e7aa0c2cfebfb7d2b5903d7d4c170238