[roll] Roll fuchsia [fx-serve-remote] Fix `fx serve-remote`
This corrects two issues that was breaking the `fx serve-remote`
workflow. First, `fx` might not be available in the current $PATH, and
so we need to use `fx-command-run` to run fx commands. Second, parsing
out the port from the `-l` was not stripping the `:` character.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1040952
Original-Revision: c567d7d2c09fa623ce76044f04ed5d229af23d7e
GitOrigin-RevId: a1aa19e48e125e59382f39611f198e01958f32d5
Change-Id: I47c19304dbf3495cddb894495cf79af032e7754b