[roll] Roll fuchsia [swd-e2e-tests] Look up ssh port in ffx
This queries ffx to figure out which ssh port we need to use in order to
communicate with the test device. This partially fixes the use of the
OTA tests with `funnel`.

Note: This unreverts http://fxrev.dev/1034317, with the following
corrections:

* Don't try to serve packages on the ssh port 22.
* Switches to `ffx target get-ssh-address`, which avoids the json
  schema change with `ffx target show`.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1040333
Original-Revision: 20ca00d1efcfae18f5d10af976dd2ba1fd698266
GitOrigin-RevId: 64cd4b6e72ea08016815e52ad7f20ca9a963e48a
Change-Id: I3ef5e293096d2f01cf1b718f1a267e6fa7a474e2