[roll] Roll fuchsia [ffx] Determine RCS state in target list (reland 2)
When running `ffx target list` when discovery is turned off
in the daemon, we discover devices locally. This CL adds the
next requirement, which is to try to connect to RCS, and get
get the product/board information.

(Added in reland) Timeout (default: 10 seconds; configurable
via "target.host_pipe_ssh_timeout") if there is no response from a
target, in which case report "unknown" for RCS and product.

(Added in reland 2) The above change only wrapped the knock
in a timeout, but the initial ssh connection was not wrapped.
This wraps the entire asynchronous query of the address in a
single timeout. Also, disable by default (enable by setting
`ffx.target-list.local-connect=true`), so we can roll this out
slowly to builders.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1038994
Original-Revision: 16c4a40cd413d4b65f30e5e3dc8deaa9df95b5b9
GitOrigin-RevId: 8a42609b8d2d671d8379a3091b78b441b9ad644b
Change-Id: I9a9b86f228d380bf6870123e9584357ce42786ad