[roll] Roll fuchsia [sysmem2] fuchsia-framebuffer to sysmem2 done
This finishes the move of fuchsia-framebuffer to sysmem2.

By renaming sysmem2.rs to sysmem.rs (instead of deleting sysmem.rs),
the line-by-line git info for sysmem.rs is preserved, and we don't
have to have the "2" in client code now that they're all moved to
sysmem2.

BUG: b/306258175
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1032396
Original-Revision: e8684889192813506c7cd9470205317b4744dba1
GitOrigin-RevId: b7fe40260869a17710879cdbc874c668aa068c8e
Change-Id: I08806f907c7de45254d578468bd73ba26fbbbe67