[roll] Roll fuchsia [serial][usb-cdc-acm] Convert to serialimpl FIDL
Implement the FIDL version of serialimpl instead of Banjo. There are a
number of behavioral changes included:
- Write() requests write the entire buffer, as mandated by the
  serialimpl FIDL protocol
- Write() requests wait for all USB requests to complete before
  replying, and surface USB errors
- Read() requests wait for some data to be received before replying, and
  surface USB errors

Also ignore USB requests that are completed with zero bytes, and a
remove a test case that was rendered obsolete as a result.

Original-Bug: 333094948
Test: usb-cdc-acm-integration-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1034096
Original-Revision: 7684733c3e1a54231f6eecbe3d2cb453095e3c6d
GitOrigin-RevId: 733b88c1f3f16b899761734907cb0564f99cd829
Change-Id: Iae3479c73bb5110a0b871a94ed2fd847c854a7b3