[roll] Roll fuchsia [swd-e2e-tests] Use the ffx that corresponds to the build
We need to use the correct version of ffx that can communicate with the
device. Initially this will be the ffx that's configured on the CLI, but
after we OTA the device we need to use the ffx from that build.

While here, this also makes sure to shut down the ffx daemon when the
test is complete, optionally support an external isolate directory, and
to delete any temporarily created isolate directories.

Original-Bug: 336897946
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/989203
Original-Revision: 46c29e20e9b062559aeda9aa1171aaa6f99a57ff
GitOrigin-RevId: 760313d9c9fc115ff34baf72577b05b02a5d994b
Change-Id: I2f74a8ac4534141782593799a083085ee0edc5a0