[roll] Roll infra/recipes [tree_status] Support using new tree status app.
This causes the tree-closer to update both the old and new tree status
apps, but uses the old status app as the source of truth for getting the
current tree status. Once the tree_name property is provided, the
tree-closer will start using the new app as the source of truth, but
until the Gerrit plugin and CV are migrated to using the new status app,
we'll have to keep both in sync, so any manually set statuses in either
app will be copied over to the other if detected by the tree-closer, and
the auto-submit builder will check both and take the latest status.

Original-Bug: 332741591
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/1041743
Original-Revision: db324826f7bd8bf77d0274b6c3a03ca0e7597973
GitOrigin-RevId: 049870e0ee9a934d537d143520ee6a9afa14c72c
Change-Id: Iffe31f59122b95462e548b4ae82bf50d1d9e8325