[roll] Roll fuchsia [netstack3] Extract EventContext to base
Extract EventContext and `FakeEventCtx` to the base crate.

We hit orphan rule/coherence problems here because of the blanket impl
on DadEvent, which is transformed in core before it's exposed to
bindings.

Solve that by adopting the same pattern used for timers: Events that
need to be converted before they're exposed to bindings do so via the
core context, and integration takes care of the type conversion.

Original-Bug: 338448790
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1041943
Original-Revision: 84279f5d9ad02c92251660be5b0d639e4f0ae8ff
GitOrigin-RevId: 28d04986aec4ada1fbd815293c3a8070e2462bce
Change-Id: Id8a66adef7a2c5e2363c29a82dfd2354f9ca15c5