[roll] Roll fuchsia [kernel][vm] Remove extraneous print
This print generates unnecessary spurious output, including when
running the kernel unittests. The correct operation when this scenario
occurs, a pinned page in a discardable VMO, is checked programatically
in the unittests and does not require this print to validate.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1041874
Original-Revision: 5eb1b87d988de11fb38da6efecd8300bf5d662ac
GitOrigin-RevId: f400715e55ec3e8d6b1f6b790ce1b9b233af4d01
Change-Id: I25084ff21313f68d7c1433d23ed5ded06a690c02