[roll] Roll fuchsia [netstack3] Extract Send/RecvFrameContext to base
Extract these common traits abstracting frame passing to the base crate,
since they're used in many modules.

We use the same pattern as `TimerHandler`/`HandeableTimer` here. Where
we have one trait that can be implemented obeying orphan rules, and
another blanket impl that can be used as trait bounds.

A first wrinkle around the test utilities is shown here, since the
blanket implementations for `SendFrameContext` in `FakeCoreCtx` cannot
be met in this world. Believe me, I tried. The solution here is to let
go of the magic and handroll the `SendableFrameMeta` trait for the fake
contexts in modules that were using it.

Original-Bug: 338448790
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1042470
Original-Revision: 5c41246fdf70cb8833aac61cd9ba5428053af275
GitOrigin-RevId: e183bd19706cb074045298c4724b630b5cf3130c
Change-Id: Ic6d632fc7b27155dd6caac87c8f3092353b46b37