[roll] Roll fuchsia [netstack3] Apply selected multicast interface on connect()
Previously selected multicast interface in SendOptions was applied only
for sendto(). It should be applied when connecting the socket in order
to select the correct route and bind the socket to the correct local
IP. This CL removes the option from SendOptions. Instead it's applied
for datagram sockets from DatagramSocketOptions.
Tests for this functionality are added in follow-up CLs.

Original-Bug: 331651296
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1041946
Original-Revision: 493091313657bb44898d5a70a8652a5acbc3c2b0
GitOrigin-RevId: 3648e68d8779c935d2b03eeb19cdddb58fb54772
Change-Id: I273dfabd112229ea991801cff408742e515e66fd