[roll] Roll fuchsia [netstack3] Extract FakeBindingsCtx to base
Core has 2 FakeBindingsCtx types: One is meant for small local unit
testing and the other is multithreading-capable and used for integration
testing. This CL extracts the small one to the base crate.

The plan is that the larger one will remain as part of testutils for the
final core crate.

As part of this extraction, clean up some of the fns in FakeBindingsCtx,
like removing unnecessary accessor for fields that are safely public or
are not meaningful helpers.

Note that to satisfy orphan rules, the `FilterBindingsTypes`
implementation has been moved to the filtering crate, showing how we'll
accumulate other trait implementations on this types as we proceed with
crating.

Original-Bug: 338448790
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1042756
Original-Revision: 614ea9418b2b2cf159282af21a822f123ed80d4a
GitOrigin-RevId: 3f569f53b459dd23e0d80276605e578f59e28c7a
Change-Id: I50c970736548523c272f2744a39edf657556dbd5