[roll] Roll fuchsia [fxtest] Build updates for e2e tests
Running e2e tests under fx test often underbuilds. This is because e2e
tests depend on components running globally on the system, but updates
to these components are not reflected in the build of just the host-side
script for the test. This CL ensures that if an e2e test is selected,
the "updates" package is rebuilt so that a subsequent publish will
result in dependencies being updated.

Note that this does not restart any components that may be running.

Original-Fixed: 305282664
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1043377
Original-Revision: e9e7578a31cb92807d3f965d9522a918d951b8e1
GitOrigin-RevId: 82563ccb5ab3dc65de00fa3816f96c625dcc6d75
Change-Id: Ice112b71128fa88a001739353fc22cf07ddf3bb6