[roll] Roll fuchsia [bazel-sdk] Allow fuchsia_package to be testonly
When a fuchsia_package is marked testonly it fails to compile because
the tasks don't get marked as testonly. This is a quick change that
allows these packages to be testonly. In the future we should change how
we create the tasks for these types of packages.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1043388
Original-Revision: 5a136d4f1c6a6d83d98e5307146adfe5e2f933a7
GitOrigin-RevId: 3032a81b52a1cf8b88133723555f242d1f8ea955
Change-Id: I3c37f8cf745fd8a93cc74c36924b1e2b661879fd