[roll] Roll fuchsia [fxfs] Add ToWeakNode
It is possible for there to be one other existing strong reference which
is dropped after the `upgrade()` call here but before the local strong
reference is dropped for not being an FxFile. If that happens we will be
dropping the last strong reference inside the lock which will trigger
the `Drop()` on the Node which will try to remove it from the cache
resulting in a deadlock.

Original-Bug: b/338634316
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1043612
Original-Revision: 58a525375e1388ae50467c88c232d78e006cccd4
GitOrigin-RevId: 5030557f106ed440788abce8f4c3d82beb527178
Change-Id: I7784e42247f2343ea2961e4aaaa2d8f03489d8c6