[roll] Roll llvm-project/libc [libc][math] fmod: clear exceptions before the test instead of after
The test has no control over the CPU state before the test runs.

This test checks whether no exception flags are set, which may not be
true at the start of the test. This used to be not a problem because the
check was broken but that was fixed in ecfb5d9951554d8bdb6a499c958f48cc35f78a88

GitOrigin-RevId: 303447617712b6c47f9c207289ba829c053c3440
Original-Revision: 49c2af031e72d9f5510003210842f51c56cf447a
Change-Id: I44327a15d61e9c5195cca9760ed6bc6b9ae76ba7