[roll] Roll fuchsia [fuchsia.io] Stop inheriting rights in Open2 when rights are not set
Previously, when rights were not set in Open2, the rights would be
inherited from the parent connection. This would lead to potential
rights escalation across remotes. This change stops the rights
inheritance when rights are not set. Now, when the rights field is not
set in an Open2 request, it would be treated as if empty rights were
requested.

Original-Bug: b/336285069
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1039512
Original-Revision: 6a0b84d086e7f415fee8662997f3afb4fecc7426
GitOrigin-RevId: 192b3bbbfcbc6b1224476242be593ee6b24afae8
Change-Id: Ib16a03a20b36ace60a4380868a16d2eacae3746e