[roll] Roll fuchsia [bt][core] clean up legacy advertising report parsing API
The `bt::hci::AdvertisingReportParser` API, as currently implemented,
resembles an iterator API whereby after calling `GetNextReport(...)`,
users can check whether or not there is another report via
`HasMoreReports(...)`. Users can check whether the parsing encountered
an error via `encountered_error()` as well. This API is very brittle,
low level, and cumbersome to use. Partially, this was because it existed
before our move to Emboss. With Emboss, we can clean up some of the APIs
and make it easier and more consistent to use this code.

This API is needlessly complex. Rather, we can do all this in one
method and return back an std::vector with all parsed reports, stopping
the parsing the first time an error is encountered. This change makes
that update.

In short, we preserve current functionality with a much cleaner, easy to
use, and easy to maintain API.

Original-Fixed: b/308500308
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1034692
Original-Revision: c0a1b0a32bde31f562f9973202af8bf3a139fe62
GitOrigin-RevId: 94dd0af5ca68de6e49159957baaf24df925c2b17
Change-Id: Iba97512feee4574143f020875035d8969f55ad84