[roll] Roll fuchsia [component_manager] removed Actions::finish
finish should not be part of the public API of the Actions struct.

Prior to this change ActionSet relied on being able to reach itself and
call finish through `ComponentInstance::lock_actions`, which is no
longer possible. To work around this, ActionSet is now held in an
Arc<Mutex<_>> so that it can hold a reference to itself and lock itself
once an action finishes running.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1041374
Original-Revision: 4cd1992c7e5b65d6967eaf1c884068b80f4c0cac
GitOrigin-RevId: ff6df1f463039d05f86cf145ce3c42456cee48cf
Change-Id: I21e7f487e118f8de9c89e7e65e08e7c2d837ecb3