[roll] Roll fuchsia [vim3][display] Add a display detection driver
Prior to this CL, the bootloader passed in driver.vim3.has_lcd variable
which was read to figure out if HDMI or MIPI_DSI display was connected.
Ideally we wanted to determine this state by probing the LCD_RESET gpio
in fuchsia itself. This CL adds a vim3-display-detect driver that does
the same and adds a display child node with the correct output type
property. With the help of bind rules, the correct drivers get started.

Original-Bug: 42076008
Test: Booted core.vim3 with Khadas TS-050 DSI display
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1043544
Original-Revision: 755274c795f9a810a3b47b10716606d2055660a2
GitOrigin-RevId: 78833419471828ae8a247362b7da3d7210d79304
Change-Id: I5d0c0c91bbf053158f55c1bb11bdffb91df9e8b6