[roll] Roll fuchsia [swd-e2e-tests] Use the ffx that corresponds to the build
This is an unrevert of http://fxrev.dev/1042292. The main change is that
we download all the blobs up-front, rather than lazily downloading the
blobs on demand.

We need to use the correct version of ffx that can communicate with the
device. Initially this will be the ffx that's configured on the CLI, but
after we OTA the device we need to use the ffx from that build.

While here, this also makes sure to shut down the ffx daemon when the
test is complete, optionally support an external isolate directory, and
to delete any temporarily created isolate directories.

Original-Bug: 336897946
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1042612
Original-Revision: aaeaeec5e9bca8d7946d249c3002d16b8e6a676f
GitOrigin-RevId: cf9a5ba41e11e87301debe13703f2adebe834d0e
Change-Id: I01468b6b3b6ebb246417f81a4884208c46841723