[roll] Roll fuchsia [debugger] RootJob should not be used in unittests
This unittest was erroneously depending on notifications from the system
state after cloning itself.

This kind of behavior is improper for a unit test situation and should
instead be left for integration testing.

Allow the respective test access to the SystemInterface and
ComponentManager internals so that the test can be performed with the
necessary set up appropriate for a unit test.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1045222
Original-Revision: 9c8e4649f895646563cd52a423bd115fd22972c6
GitOrigin-RevId: ae595156d21b376ec87d81d0525ac856d544fe29
Change-Id: I268e96eddbb8ff28d8135b1473105f6132e07035