[roll] Roll fuchsia [kernel][vcpu] Fix crash caused by missing chain lock transaction
~Vcpu (arm64) had a bug where in some cases (depending on who wins a
race) it would attempt to acquire a chain lock without an active
transaction.  See b/339326078 for details.

Update the arm64 ~Vcpu to create a transaction before attempting to
acquire the thread's lock.

Add a DEBUG_ASSERT to ActiveRef to make this kind of problem easier to
debug in the future.

Original-Fixed: 339326078
Original-Fixed: 339331212
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1045040
Original-Revision: ec6abd008ac6349526506e5852d2516932b634da
GitOrigin-RevId: eda73622ef39ddb68b44bd0aa233cfa208884abc
Change-Id: Idafe418dcb5e4abce796acfc3295ab6927dda4a6