[roll] Roll fuchsia [uart] Fix blocking option not being surfaced into the waiter.
There are two flavors of dputs one that will never block, and one
that may block. I mistakenly mapped may block to the case were it would
not block when debuglog is disabled.

Note that the splitting of the flavors is correct, just the mapping for
the newer drivers was incorrect. The Waiter needs to, first check which
flavor it is, and then check whether its allowed to block.

The fix is just forwarding a template parameter to Waiter from the Write
call.

Fixes: 336859048
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1036197
Original-Revision: 8721f6eb9e0614d3da0077235a087dc06d327bac
GitOrigin-RevId: 09412dbde1759e495aba23a83f880019f68e7d9b
Change-Id: I5f617f5a9ed2b70fba16c682f48589d4527f626e