[roll] Roll fuchsia [netstack3] Rename FakeEventDispatcherBuilder
`EventDispatcher` is a very old name for the bindings contexts that is
no longer used, this is an artifact from a distant past.

The builder is still used to create `FakeCtx`s for tests, so renamed it
to `FakeCtxBuilder` instead.

Original-Bug: 338448790
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1045596
Original-Revision: 661526e53dc1c0f2867a2136b1318e8f91a30a45
GitOrigin-RevId: bbe58843a023684a09b88658321fb0b3c7d0dd92
Change-Id: I1edc82397437e2206d8f8c24f0c19ffbe29a9b08