[roll] Roll fuchsia [kernel][chainlock] Don't touch lock contents after release.
Make absolutely sure that we never touch the internals of a ChainLock
after its internal state has been set to Unlocked.  This aligns the
ChainLock implementation with special requirements defined for
std::mutex; basically that as soon as a lock is free and can be
obtained by someone else, the code releasing the lock has to consider
the object to be dead and cannot touch any internal lock state.

Prior to this, we were reading the ChainLock's contention mask
immediately after setting its state to Unlocked, which sets up UAF
potential.  See comments in the CL and http://fxbug.dev/339437199 for
more details.

Original-Fixed: 339437199
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1045611
Original-Revision: b8d4ecbbacd37d63617b365a3768456029ad1944
GitOrigin-RevId: 727135b22711c845527658efbb4bd0db7907b10e
Change-Id: Id57b354ae2b749070205acacf659d1e621d38325