[roll] Roll fuchsia [ffx][fho] Change verify_schema to associated func.
In VerifiedMachineWriter there is a function that takes `&self` that is
unused. As shown in this CL, requiring this makes it difficult to test
things when a writer is passed by value into a function.

This changes the verify_schema function to be an associated function
instead of a method to rectify this.

Original-Bug: b/339725104
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1046093
Original-Revision: 0514351ab8e6d2b82828bc0b1cc3e5751c5f0154
GitOrigin-RevId: 181b536f73395111ffd00b8df92ee4ed8597ea90
Change-Id: Icbed8b6955f65a37ebcdf5962bfd11eacace3aa0