[roll] Roll fuchsia [component_manager] Fix shutdown panic bug
It's possible to cause shutdown to panic by constructing a cycle with
dynamic offers. Specifically, this is possible if there is a service
offer from a collection, followed by a chain of dynamic offers that
routes back to this collection.

To make this impossible, detect these cycles at component creation time.

Original-Fixed: 297403341
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1030252
Original-Revision: 9439d55d466bb0354bdfcfa704ce9399a8de0959
GitOrigin-RevId: 5f42b4dbd7821e5ad6c4be99f6e9e442590a9478
Change-Id: Ic9fdc8872b35e4ecb96eb79f739ce387f3d4a1dd