[roll] Roll fuchsia [kernel][chainlock] Don't evaluate types during if constexpr (false)
Make some of the internal chainlock methods templated in order to
prevent type expansion from happening in the false half of an `if
constexpr` statement.  If an `if constexpr` occurs in a non-templated
entity, both halves of the statement are fully checked, which is
apparently something that has always been the rule, but which is just
now starting to be enforced in compilers.

See the (large) new comment included in the CL for details.

Original-Fixed: 339652126
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1046152
Original-Revision: 50a6426de756023787a4d8974c7b74b003a7935d
GitOrigin-RevId: 838810271711485a7695f0a27f10e82071458eb5
Change-Id: I2cdae611d0248d97e34f16fa0f093273c1eddf05