[roll] Roll fuchsia [ffx] Support ffx --log-output <file/stdio> ...
Add an option to direct the output to a particular destination.
If the destination file is "-" or "stdout", send the output to stdout.
If it is "stderr", send to stderr.

This reinstates the functionality (lost in fxr/1042050) that daemon
output should go into ffx.daemon.log instead. (It also makes it so
that `ffx daemon start` produces output on stdout, retaining
compatibility, with the previous behavior, since this is necessary
for integrators who manage the daemon lifecycles themselves.)

Original-Bug: b/338451284
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1045226
Original-Revision: 3f61f5b6eed9e3766b55fef018a4771f3ef95721
GitOrigin-RevId: 6a3c40fd2eac327d0b1ab0bc8b66198f67ec4f62
Change-Id: Iba1d37dd5badc5a8b912f27fd3925d3e899ce033