[roll] Roll fuchsia [power][storage] Create system integration test
This change adds the system integration directory with a test created
for storage. This test includes a custom product and board that infra
will flash onto a vim3 device and then run the test in isolation. The
custom product is a bootstrap level product, so it will be run on
bringup builders.

Original-Bug: 335526421
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1042046
Original-Revision: 64d2b1b3a1b5576668e8abab0e20ce2f131c0c61
GitOrigin-RevId: fa4c8253377f9f7a9a92447cf146de1ae3521fff
Change-Id: Id9884befef4fe1a162772ebe733720941f665148