[roll] Roll fuchsia [starnix] Add ffx command to suspend container
This command currently suspends all running Starnix containers by
suspending each process in the container.

It does not suspend processes that may be logically associated with a
container if they are not in the same job.

In order to make it simple to suspend a single container, we could add
suspend functionality to fuchsia.starnix.container.Controller. This
would allow clients to connect to the container they wanted to suspend
directly. However, the suspend operation would still need to be
performed by a separate component, outside of the container's job. This
means that we would need to rework the division of labor between the
starnix_kernel and starnix_runner (mainly which component serves protocols
on behalf of the container).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1043153
Original-Revision: e647d207311fb6be144bea39e6d1563ede193931
GitOrigin-RevId: 2436019525de118814f03e88eec13b6733c41dd9
Change-Id: If39fd2f433d163f0c76528f0254ccffdbb763aff