[roll] Roll fuchsia [netstack3] Introduce TransportPacketMut
Introduce variants of the transport packet methods/traits in the
filtering crate that allow for packet modification (not just
inspection). This includes MaybeTransportPacketMut and
TransportPacketMut. For now, the only methods on the latter are for
updating the pseudo-header in the transport-layer checksum, which will
be necessary to allow updating the source/destination IP addresses in
the IP header.

For packet builders, updating the pseudo-header is simple because the
packet has not yet been serialized and the checksum has yet to be
calculated. For incoming packets that are partially parsed, introduce
the ParsedTransportHeaderMut type, which holds a parsed transport-layer
packet allowing for in-place rewriting of its fields.

Original-Bug: 321013529
Test: netstack3-filter-test
      netstack3-core-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1044975
Original-Revision: cb37d625f177b7fce36112060d9dd9eca3e0e854
GitOrigin-RevId: 2550cd49fdf47cbc3012afb70512247d64c4b4ab
Change-Id: I8bf8ec345fcc0694cfec3dee246a877ef698a747