[roll] Roll fuchsia [netstack/conntrack] Only collect connections once
By only checking if a connection is expired when we're looking at it
through the original tuple, we cut the size of to_remove in half. It
also means we can add in asserts that we actually did remove the
connections from the table.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1046619
Original-Revision: 7cee93f3312716189a360d6cd97b8f595b87e6c6
GitOrigin-RevId: a913c8a275c2943957a1dd2d459a78dd9fd8893f
Change-Id: Idefd948e84f652f6614b481122b1e81911aba553