[roll] Roll fuchsia [component_manager] concurrent start and stop action may not cause event
The model::actions::start::tests::concurrent_start_stop test had an
assumption that if a start and stop action were run concurrently, at
least one of them would emit an event. This is an incorrect assumption
because if the start action is started first and then immediately
cancelled because of the stop action, the stop action will then not emit
a stop event because it only does so if the component had been started.

fxrev.dev/1042096 slightly changes the timing of actions being run, and
causes the above flake to occur at significantly higher rates.

To resolve the issue, the event expectations have been removed from the
test.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1046372
Original-Revision: 99ff01beb8976519b58530b85b69fdf35d8d1a39
GitOrigin-RevId: 49cf6825c402713c27ffc88b6c67b0d83acf55e0
Change-Id: I618ba5a9f9a63d3ff0a9bd78e15bb6356c18aa6d