[roll] Roll fuchsia [packet-formats] No pseudo header in ICMPv4 checksum
ICMPv4 packets produced by packet-formats do not include the IP pseudo
header in their transport-layer checksum calculation, because the IPv4
header always includes a checksum. Fix the pseudo header checksum update
logic to correctly do nothing for ICMPv4.

The existing test did not catch this because it was swapping the source
and destination IP addresses, resulting in a checksum that was the same
as if those addresses had never been included. Fix the test to use
different IP addresses for both source and destination such that it
catches this bug.

Original-Bug: 321013529
Test: netstack3-filter-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1048072
Original-Revision: 0330c7696ccd4141570f33dd9f92c5861d62afb4
GitOrigin-RevId: f70d782281edd8fbcd027ad4dd17857878f732b6
Change-Id: I710a491fe9c8932f108847191e4e6261f5ce770b