[roll] Roll fuchsia [bazel_sdk] Make fuchsia_transition() idempotent.
Due to the way Bazel works, the transition function can be
called multiple times in a chain of dependencies, and should not
change the content of the input build configuration if it is already
the result of a previous fuchsia_transition() call.

This fixes an issues where `-ffuchsia-api-level=XXX` was appended
to the `copt` list on each call, resulting in a different build
configuration instance on each call.

+ Add @fuchsia_common//:transitition_utils.bzl to export new
  functions `set_command_line_option_value()` and
  `remove_command_line_option_values()`, as well as appropriate
  unit-tests, which can be run with:

  ```
  fx bazel test @prebuilt_clang//:test_suite
  ```

+ Remove obsolete key from NATIVE_CPU_ALIASES, the
  associated value (aarch64) was wrong anyway.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1047975
Original-Revision: dc66b6dc4bd90b8f44f537c44b17c8a19c8b914a
GitOrigin-RevId: 77767b5695e1e810e6de56cc3c762ddacfb1045e
Change-Id: I2d1cf67c5c502d82cd3bf3bfcaa21d8923310b50