[roll] Roll fuchsia [netstack3] Prepare loopback for crate split
Split everything that will be in the integration core crate into a
separate file. We had to mark tx and rx queue metadata separately to
satisfy the fact that we need different OrderedLockAccess
implementations for the tx and rx queue state. The pattern will need to
be applied again to Ethernet and IP devices in a follow up CL.

This is still not fully clean, the blanket implementation of Receive and
Transmit queues BindingsContext traits will not work post crate split.
We'll need to massage the traits there.

Original-Bug: 338448926
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1048414
Original-Revision: f1a781e2bf743cf7fcdac8d474945d2395df5840
GitOrigin-RevId: e4ae67a5f9da2ddc234d448ccca44371c868ab32
Change-Id: I83b0e15e47ab9d73ff79e34b28eb90ff4af44dc3