[roll] Roll fuchsia [fx] Updating remote scripts to use ~/ssh keys
This removes the consideration of the in-tree ssh key when checking
ssh keys between the local and remote computers.

This is being done to align in-tree and out of tree ssh keys by using
`ffx config get ssh.(priv|pub)` as the source of truth for the location
of the keys both in and out of tree.

Original-Bug: b/339034335
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1045213
Original-Revision: 88b718f13d7fb27a4507fadff9ff97cefb760188
GitOrigin-RevId: 64b313f07675ce2761973e4f38f75e052097abda
Change-Id: I997c40023873120ea10959cdd4720c461b4a6919