[roll] Roll fuchsia [netstack3] Refactor IP dispatch to use packets
... rather than raw buffers, which enables a follow up CL to deliver the
packet to raw IP sockets.

Because `Ipv{4,6}Packet` contains a reference to the underlying buffer,
the packet is invalidated whenever the underlying buffer is moved (due
to Rust's move semantics). This happens when passing the buffer as an
argument to a function. By passing the packet rather than the buffer,
the packet remains valid, and can be used both for raw IP socket
delivery and delivery to the transport layer.

Note, that one drawback of this approach is that the transport layer no
longer takes ownership over the underlying buffer. This may have been
useful for optimizations that want to reuse the backing allocation, but
such optimizations are too far down the road to realistically plan
for.

Original-Bug: 337825157
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1048352
Original-Revision: 3e5872c54dbf1ff3d9d8d75467e57866d703030e
GitOrigin-RevId: 0cba27ffee03cf4569d3dfad2cbb88df5914be79
Change-Id: I6a1c7969ade5e5673121caf4f69ea42dca07e06b