[roll] Roll fuchsia [display][coordinator] Do not keep fdf::Arena in EngineDriverClient.
fdf::Arena and fidl::Arena are designed to be for ephemeral use only.
Once a piece of memory is allocated in an Arena, it's never recycled
until the Arena is destroyed.

Thus, the fdf::Arena should never be kept as a member variable in
display coordinator EngineDriverClient, otherwise it will eventually
cause an OOM.

This change replaces the arena_ member variable with local fdf::Arena
created within the FIDL call scope; this is the common practice in
driver code across the Fuchsia repository. Note that each fdf::Arena
creation calls the fdf_arena_create() function, which will allocate a
new Arena on the heap.

Test: "display-tool squares" on core.vim3
Original-Bug: 325474586
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1046639
Original-Revision: 9c385c48c446b20e5f601888366585ab02e2dd19
GitOrigin-RevId: 8893326c92198d95e811a90af58b3fc34537cafc
Change-Id: I673e1c8adb734ac6975c7c63d09e923591c961ee