[roll] Roll fuchsia [audio][drivers] Add ProcessingDelay property
Add `processing_delay` to ElementState and deprecate `latency`.
Although the `latency` union is a superset of processing_delay,
only the duration aspect of `latency` is ergonomically optimal
(the other option is to specify this in frames), so in practice
it is simpler to just be a duration, named ..._delay like the
others (external_delay, turn_on_delay, etc).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1046071
Original-Revision: 6cbfff71b0e34dbff3745466b8e386f6aa93574d
GitOrigin-RevId: c8810d132cd156d04472b3690d16c33f751a9c4f
Change-Id: Id5aacc709362c4d192eec6a7b74dbc1764d0e72f