[roll] Roll fuchsia [component_manager] pass program output into dictionary construction
Passes the program output dictionary directly into dictionary
construction instead of reaching into the component for it. This is
necessary because we won't be able to access component internals so
easily once this code is moved to the routing library.

This change would have introduced reference cycles between dictionaries,
so a WeakDict type is introduced. A WeakDict can be upgraded into a
Dict. This was not an issue before because all dictionaries that wanted
to route into a program output dictionary would hold a weak reference to
the component, and upon receiving a routing request they would then
upgrade that reference and grab the component's program output
dictionary from there.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1044832
Original-Revision: 1eafc18fbffacc78f682a3c62e921a840cc74db6
GitOrigin-RevId: d60ba5b72f4f699e794ecd78ab7d8e6c520f02c2
Change-Id: I1a9d8cc42ca271b8895a0d623a1bbdab81651143