[roll] Roll fuchsia [fidl][ui] Mark brightness protocols as server=platform
As part of RFC-0241 we are marking protocols whose servers or clients
are only supposed to be in platform or external components as such.

This marks all of the fuchsia.ui.brightness protocols as allowing
clients in platform or external components but only allowing servers in
platform components.

Original-Bug: 335446415
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1049682
Original-Revision: e040a5356511e561b657cb2e22935da3bf32178a
GitOrigin-RevId: b8c523874903d3b84ab9ad712c262fa38b0a41e0
Change-Id: I7e8277729423be347468aaa48b6d47cf584233e0