[roll] Roll fuchsia [starnix] Handle state explosion due to branches
The previous verification algorithm will explore each possible branch of
the program. This induced an algorithm that was exponential in the
number of conditional branches.

To solve this issue, this CL introduces an algorithm inspired by the
description of the linux kernel algorithm in https://docs.kernel.org/bpf/verifier.html

The verifier use liveness information to broaden any state it can verify
will correctly terminate. It then keep track of all these states and
when a new state is produced, it checks whether it can already know it
will terminate, and in that case, skip the computation.

This CL also adds a test that exhibit the exponential behaviour of the
previous implementation.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1039632
Original-Revision: 0b2b14e6e2900dcf55a7a52379f0469a6a656c17
GitOrigin-RevId: 1d8b51d219e8c4a61ac830187a386ce7b0173571
Change-Id: Idd83c38049be285bfa8c822c657e730c0b947ba1