[roll] Roll fuchsia [bt][a2dp] Fix compatible endpoints not matching
Update the logic for matching compatible codecs (during start) to
include all endpoints which we can negotiate a compatible set of codecs
for.  We continue to be stringent when setting our configuration.

Add negotiation to the start process to use the highest quality codec
parameters that are compatible.

Original-Fixed: 339159944
Test: added unit tests
Test: manual test with previously failing headset
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1048244
Original-Revision: 2e97d8350fdcb448fca8db73363c538bbd2c18d1
GitOrigin-RevId: d3750f28bb1f522fd4c51795acf0ee6c5801b669
Change-Id: I47a54ec769c4940dbd2f5c60671201e320689a16