[roll] Roll fuchsia [vim3-devicetree] Fix network connectivity issue
There were two issues which was leading for the vim3-devicetree based
devices to be unreachable. One was the netsvc_interface path, which
differs between vim3 and vim3-devicetree (on vim3-devicetree, the path
name is derived from the node name and changing that to match that of
vim3 is a major change). The other issue was passing of the serial
number metadata was missing for usb peripheral on vim3 devicetree.

Original-Fixed: 340578181
Test: Checked that the device was accessible through fx shell.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1048094
Original-Revision: 695acba7fdab5902b2446e504c21bc720320ac1d
GitOrigin-RevId: ab4ea6e0d7b61d380cf0acc7842e008f73d17792
Change-Id: Ib99a2d93e6d474e021855fc411fd4da64e30705b