[roll] Roll fuchsia [kernel_zbi_test] Propagate device_type to cuckoo test
The test should only be executed in the device types set in the
zbi_test. This were not forwarded since it expected environment,
but boot_test expects device_types. So we translate device types to
environments.

Reenable cuckoo test, but given it will not be executed outside of nuc7,
and coverage is only QEMU bots for the time being, then the overall
effect is the same.

A separate issue is HPET reliability in QEMU.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1049718
Original-Revision: 04d74778238963f4f75897869a6e8628a73401e0
GitOrigin-RevId: 10ef05da10424211dd8dacc56e1f50f841749b7b
Change-Id: I99a3b31a1462abfd790b4aba6bbdf0b310286831