[roll] Roll fuchsia [fsverity] Set the digest_algorithm/size output fields in measure_verity
In its old implementation, fsverity treats the digest_algorithm field of
the fsverity_digest struct as an input field but it is actually only an
output field. This change sets both the digest_algorithm and digest_size
of the fsverity_digest struct that is returned to the user. This should
also fix the begonia boot regression bug attached to this CL, which was
caused by the EINVAL error returned from a check that treated the
digest_algorithm as an input field.

Original-Bug: b/340066061
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1049193
Original-Revision: d0ddd02b8224c662c2b24ca7a2331864ef7f2ae3
GitOrigin-RevId: 3af6b2bfebee719d70b4a3dd620bb133d03d50f9
Change-Id: Ibd92947e5a723fd21fb49e08b425038fd9538adc