[roll] Roll fuchsia [ld] Support preloaded modules in RuntimeDynamicLinker::Init
This allows an InitModule to indicate that a module should be
consulted for dynamic linking but is already present in the
address space and does not need to be loaded.

The new ld::RuntimeDynamicLinker::PreloadedImplicit convenience
function turns the initial modules from one dynamic linking
session into preloaded implicit modules for a new session.

This is demonstrated in a test where a main executable with a
DT_SONAME and dependencies of its own is launched via remote
dynamic linking, and then later a second root module is loaded
into its own dynamic linking namespace using the main executable
and its vDSO as preloaded implicit modules to satisfy a DT_NEEDED
on that SONAME.  This approximates the intended case of a driver
host process and a separate driver loaded into it by a service.

Original-Bug: 326524302
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1047433
Original-Revision: ef7fd1a8dea9fb41e633a33de4d7c35f1d6d9f84
GitOrigin-RevId: 796d58cbc8a91c04f2ee302cc89e442e101ebf3a
Change-Id: I24a7bfb8401ac7bd8f572e08e8ba9f1b2f5b12e3