[roll] Roll fuchsia [fidl][fuchsia.diagnostics] platform-only APIs
As part of RFC-0241 we are marking protocols whose servers or clients
are only supposed to be in platform or external components as such.

LogSettings is not allowed to have clients or servers outside of the
platform.

ArchiveAccessor can have clients outside the platform but servers can
only be in the platform.

Original-Bug: 335446415
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1049676
Original-Revision: afe2b2a8d97da9a10dc0bbabe786dfc8752b13f8
GitOrigin-RevId: dca2c7a835ac385e5c61c3b05dc528e1d0511648
Change-Id: Idb51a514f7c3a372790c5c97651922659bfe36ba