[roll] Roll fuchsia [starnix/usercopy] Don't use hermetic code blobs
...for copy/zero routines now that previous changes introduced the
routines in assembly.

Given the complexity of the routines, this change continues to build the
copy/zero routines from their C++ sources in tests. This is to, maintain
a valid reference in tree that we can use to update our assembly with.

Original-Fixed: b/338636637
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1045757
Original-Revision: 04ea7b2921d977c9afc11c80fffb73ecd7ee82ea
GitOrigin-RevId: f4bf749f28d7e9786974e67a5cfc758318aec290
Change-Id: I7b9ec695d67d5afef45911565401749df5d28404