[roll] Roll infra/recipes [auto_roller] Rename CQResult -> PresubmitResult
Since this module is sometimes used for submit CLs in repos that don't
use CQ, and supports monitoring other presubmit systems, we should
generalize some of the terminology.

There are other places in this module where the terminology should
probably also be updated, but it's not worth putting in the time to do a
thorough sweep at the moment.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/1050472
Original-Revision: cf8de90b42ed7a20c84d7c0570b4c5a731feac94
GitOrigin-RevId: 2cf2e300b85d5aeaa86bdf4377eeb92752f0dc6a
Change-Id: Ic80af45a967f5ed688d9b9b865863981bcc63adb