[roll] Roll fuchsia [fidl][kernel] Mark protocols as server=platform
As part of RFC-0241 we are marking protocols whose servers or clients
are only supposed to be in platform or extenal components as such.

This marks all of the fuchsia.kernel protocols as allowing clients in
platform or external components but only allowing servers in platform
components.

Original-Bug: 335446415
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1046054
Original-Revision: 968ed5c4386d539a0ffc4633ed01c08584460d73
GitOrigin-RevId: b4c996d964d149ccf40313db39780e8a9a596b5d
Change-Id: I6f5d93658403d0ab785c92e9d95ec5eb63364391