[roll] Roll fuchsia [fidl][settings] Mark protocols as server=platform
As part of RFC-0241 we are marking protocols whose servers or clients
are only supposed to be in platform or extenal components as such.

This marks all of the fuchsia.settings protocols as allowing clients in
platform or external components but only allowing servers in platform
components.

Original-Bug: 335446415
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1046057
Original-Revision: a1a1085d97a836c924a27cd46b11ab8856de8178
GitOrigin-RevId: 9a2c207d9be2ad84d713f7f9c11211c78a733221
Change-Id: I53ee1eef69788481219f9a7e6c34f85064afaef0