[roll] Roll fuchsia [netstack3/conntrack] Fix self-connected sockets
Conntrack has tacitly assumed that for all connections, the original and
reply tuples will be different, but this is not the case with
self-connected sockets.

There are a couple of fixes required here. The primary one is to track
the number of connections separately from the HashMap. Then, we also
need to insert a few ifs to fence off a couple asserts.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1050513
Original-Revision: a6edd1f7bf1ca071abdbda167799b5964c2dc59f
GitOrigin-RevId: d2d471912e59ae1119252e05d2e6937e64d1df34
Change-Id: I42ee50935bdbdbf6b470b1eb1eaac9585f11cbae