[roll] Roll fuchsia [netstack3] Reduce freestanding fns in socket mod
All the helpers in the netstack3_core::socket module will be moved to
the base crate. Tying the freestanding functions to well named types
makes them more ergonomic in the long run and sidesteps the desire to
import freestanding fns.

While at it, noticed that the TCP module was attempting to upgrade a
bound device ID in an error path which should not be done so massage the
helpers in socket and TCP to correct that.

Original-Bug: 338448926
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1050598
Original-Revision: 0694a828e3dbf83283914ba00bc492990ab913f5
GitOrigin-RevId: 9cb39bd3cd400ed3d7a3222be7fded5cf5a7c6bf
Change-Id: Ic1a76422b7e13406f3e3f9c7272fffa57158c87e