[roll] Roll fuchsia [netstack3] Introduce SocketZonedAddrExt
Instead of a freestanding function in the transport module, move
resolve_addr_with_device to an extension trait on ZonedAddr so we don't
need to import freestanding functions, and define it in the socket
module which will be extracted to the base crate.

This allows TCP and UDP to use the extension directly from the base
crate without the need for a base transport module.

In preparation for moving socket, move EitherDeviceId to the base crate
so the new extension trait can reference it.

Original-Bug: 338448926
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1050599
Original-Revision: ee2d1e6f225dc7546ebea7d620a957d2d16f8170
GitOrigin-RevId: 80f76724678738e6934af06cf34d762d36a46f0a
Change-Id: I89a13730b89c59bbc5dceb21fc4a66df5467c84c