[roll] Roll fuchsia [wlan] Handle framework errors in fidl_bridge
ForwardResult now explicitly checks for framework errors. ForwardResult
will now close the completer if the framework error is due to channel
closure.

For errors that do not result in channel closure, ForwardResult sends
the framework error to the caller as a domain error if possible.
Otherwise, it logs and ignores the error and replies to the caller with
a default-initialized response if necessary.

Unit tests have been added to show that channel closure propagates
through a bridge server using ForwardResult.

Test: fx test drivers-only/fullmac fidl_bridge_test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1043192
Original-Revision: 2779301253217c22ad4cf2db640edf385fb90689
GitOrigin-RevId: ba15a2e1690d184c249837090fa44a3569fbc748
Change-Id: I214b762c3323b8987736f86a1940e14b0144a5c2