[roll] Roll fuchsia [netstack3] Add IpPacket::set_{src,dst}_addr
Add methods to IpPacket that allow modifying the source and destination
IP addresses. For IPv4 packets, this requires updating the checksum, and
for all IP packets that contain a valid transport header, it involves
updating the transport-layer checksum to reflect the new address in the
pseudo-header.

For ForwardedPacket, which wraps an entire IP packet which has had the
IP header parsed to allow for buffer reuse, updating an IP address in
the packet header involves reparsing the IP header so it can be modified
in place, and then updating the checksum(s) as necessary.

Original-Bug: 321013529
Test: netstack3-filter-test
      netstack3-core-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1044953
Original-Revision: 3cfa3b22edfb5391171eed30cc6804e0d7feeb8b
GitOrigin-RevId: 25b311f43ac6ccdad9f11cc8f7049af91a409d90
Change-Id: Id5e65b35f1f07338205cf7b1710c842556488bab