[roll] Roll fuchsia [netstack3] Add TransportPacketMut::set_{src,dst}_port
Add methods that allow changing the source and destination port in the
transport header of an IP packet. For incoming packets this requires
updating the checksum in place as well.

Note that for ICMP packets, it continues to be the case that only echo
requests and replies support transport-layer inspection and
modification. Just as an ICMP echo request encodes its ID as the source
port and a reply encodes its ID as the destination port, only those same
ports can be rewritten.

Original-Fixed: 321013529
Test: netstack3-filter-test
      netstack3-core-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1046012
Original-Revision: 680361f71b342f8c426d2a524fc9fb199991b71c
GitOrigin-RevId: d1bba9563fe0922a4a50ca3d45d6ae392d8ac69f
Change-Id: Id65752e24725398fecaf416ff0ce2046d46538dc