[roll] Roll fuchsia [ffx][config]Making "file" processing return 1 string
This changes the `--process file` handling in `ffx config get`
to return the first path configured that exists. Before this
change, if the config value was an array, it would return a
single value if only 1 of the entries in the list existed, but
if more than one existed, it would return an array, making it
very difficult to properly handle.

This change makes it always return a single entry.

Original-Bug: b/341777920
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1051241
Original-Revision: 746af24b49066e4e15b7b32e8a159fe787a8f8a1
GitOrigin-RevId: 8e5ec3daad4234a5f42d0ce445d53f9706522d09
Change-Id: I9db0420748cb3c81989648229e21aa10d4f03d06