[roll] Roll fuchsia [starnix][selinux] Replace has_permission with has_permissions
Replace the generic `IntoIterator` argument to `has_permissions()`
with a slice argument, to simplify the implementation and allow
Rust to take advantage of the fact that sets of permissions will
always be expressed as static arrays.

`has_permission()` is removed, since `has_permissions()` is
expected to be sufficiently simple to be fast-enough for single-
permission checks.

Original-Bug: 335397745
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1028973
Original-Revision: 2b87afe1fbb954f165c2850daf8cfb5d83e2f74c
GitOrigin-RevId: b0ade7f900055365150c8e0fbed1421aa0e686a9
Change-Id: Ie15e7ad6e99a39ef21367e3017c218864f7dd71e