[roll] Roll fuchsia [ramdisk] remove GetName support
For some reason, ramdisks supported the Partition.GetName fidl call,
returning their name as "ramdisk-x". This makes sense in a vacuum, but
since ramdisks are almost always used in place of real disks, they
should act like real disks wherever possible, and real disks don't
support this call, it is generally reserved for block devices served out
of partition maps.

This is required for a change to fshost which checks the result of
GetName when matching a device for fvm or fxblob. The change succeeds on
real devices but the integration tests failed because ramdisks didn't
follow the same pattern.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1048392
Original-Revision: 335d8d344088f540913d96c4bcd26caf5be1fbbf
GitOrigin-RevId: 17608a3fa9e815944cb2425f555a8fa50f7c0d2b
Change-Id: I37a6324a12f722b71e3af3abcacd7ac786b497ca