[roll] Roll fuchsia [fshost] fvm/fxblob opportunistically check device label
Have the matchers for fvm and fxblob try to check the device label, if
it successfully gets a label they make sure it's the right one. There
are some devices we run filesystems on that don't support labels, so if
there are errors then we ignore them and continue checking the content
type.

Original-Fixed: b/293948720
Test: fx test fshost-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1045275
Original-Revision: 0b134c9b59315e522ef9e026bc0625f230adfb42
GitOrigin-RevId: 650c7418725a6e24c01e22cd87018e051160c191
Change-Id: I6dcff33852affedf1dc13970effa5a27c1fb98ea