[roll] Roll fuchsia [inspect][graph] Improve initial metadata event recording
Instead of iterating twice over all the metadata, we can construct the
event node as we record the metadata and then reparent it when creating
the event. This removes the need of requiring that the iterator is
clonable as well and simplifies the implementation and performance of
the follow-up change which brings nested metadata support.

Original-Bug: 340342858
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1050233
Original-Revision: b6fe9c7ff92cd9cf1bc66f3f77d4b168b3b0b4d0
GitOrigin-RevId: 58a5af641fbb6be9b880fda1491fa50bf6baede9
Change-Id: I563827886650dfac84f46d63f5156522df491aa5