[roll] Roll fuchsia [tee] Add manager component
This adds a TA Manager component which is responsible for instantiating
TAs according to the configured policy and monitoring them. Once we add
the ability to route capabilities to specific TAs the manager will be
responsible for this as well. The first version of the TA contains
configuration parsing logic to determine which TAs exist and what
implementation we want to use for each UUID.

This iteration has support for knowing if a TA supports multiple
instances but it doesn't use this information yet. As such, it does not
correctly handle multiple connections to the same UUID.

Original-Bug: b/332956721

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1042480
Original-Revision: dc7f078e6b72701dda5dcd895a8f708e06ad6ba6
GitOrigin-RevId: 02155b8fef606a3981b56a375e34d8081d15582a
Change-Id: I2c21e3534000fd70ca8a466606b24257d2525d8b