[roll] Roll fuchsia [assembly] Properly handle @gn_targets path when exporting packages from Bazel to Ninja.
When a package is assembled with Bazel and includes inputs from the Ninja
build directory, its assembly manifest will contain paths that reference
the @gn_targets repository, or symlinks that point to files inside the
latter. The files in @gn_targets are themselves symlinks pointing to the
Ninja output directory.

Because the content of @gn_targets is regenerated every time a different
GN bazel_action() is built, these paths can point to stale symlinks
when they are used outside of the Bazel invocation that generated
them.

This can happen with the following sequence:

1) A package is generated by GN, even if Bazel assembly is enabled.
   This happens for example for any fuchsia_shell_package() target
   such as `//third_party/sbase:basename_pkg.pkg`.

2) The assembly_input_bundle() GN target is built, which reference
   the package in 1.

On the Bazel side:

  `//build/bazel/assembly/assembly_input_bundle:platform_eng`,
  a `platform_aibs()` target, depends on:

  `//build/bazel/assembly/assembly_input_bundle:core_realm_eng`,
  a `fuchsia_platform_artifacts()` target, which depends on

  `@gn_targets//bundles/assembly:core_realm_eng` which exposes
  the

3) //build/bazel/assembly/assembly_input_bundle:platform_eng is
   a Bazel platform_aibs() target, which references the

   The corresponding manifest file and content are stored in the
   Bazel output_base.

Fortunately, their format is fixed intentionally to allow the extraction
of the corresponding Ninja output artifact easily.

Original-Bug: 42082672
Original-Fixed: 340667657
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1051952
Original-Revision: e68e51c7d27efb2dd26893b55c856fd2eae8d464
GitOrigin-RevId: 9cb46b7dd3120a2392aafd0354272a3db94403e5
Change-Id: Ide02368d08312cfdbc3e02d4bfbdf928dbbc7d4e