[roll] Roll fuchsia [netstack3] Add NatContext
Define an execution context for performing NAT, which currently includes
one method that provides the ability to get the best local address for
communicating with a particular remote, over a particular interface.
This is useful currently for Redirect NAT in picking a destination
address for an incoming packet, but it will also be useful in the future
for Masquerade NAT (a special case of SNAT), where it will conversely be
used to select an appropriate source address for an outgoing packet.

Original-Bug: 333419001
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1049953
Original-Revision: 5742945fb1b1a3201a6069927ee2965710f09bb4
GitOrigin-RevId: 0c092949f04640dba46aba6a8462575c286521dc
Change-Id: Ic2223ead689b14ab03060086f6d1ed070d2eb082