[roll] Roll fuchsia [bazel_sdk] Fix structured configs with 'config' section in CML file
The GN version of fuchsia_component_manifest has a hardcoded bit
of code that references structured configs. This copies that logic
over to the bazel template: http://shortn/_DWGNHmiTHD

Also fixed an issue that caused the outer brackets of the values dict
to be dropped when it was written to a file.

This also allows boolean values to be written as boolean values
directly, not strings of "true" or "false".

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1052052
Original-Revision: eeb1245658de60a23d103146dc39fa7233af1556
GitOrigin-RevId: d1c55612cb4aa622a172371d23068fac29fac696
Change-Id: I1c414a428fbe8b9fef027c379d675c5b775452ec