[roll] Roll fuchsia [kernel][chainlock] Be sure to watch the proper CPU when backing off.
When encountering a ChainLock conflict and backing off, remember which
CPU we had the conflict on before unregistering our active CLT and
waiting for the lock we wanted to be released.  Failure to do this can
lead to a situation where we migrate to a different CPU which _just
happens_ to have the same conflict ID as the CPU we just came from.
If we stay on this CPU forever, and nothing ever generates a new
conflict on our new CPU, we will just stay stuck forever.

See the detailed write-up in b/341370301 if you are interested in the
precise sequence.

Original-Fixed: 341370301
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1052096
Original-Revision: 48ba01297af58e2e4f08712dc757c9ac9c9637ad
GitOrigin-RevId: 385b6f9acddbfa7edc30de881bc30828843440e1
Change-Id: Ib2d53c4823ffabd834caef33be01b3d50b024f14