[roll] Roll fuchsia [vim3-devicetree] Add missing NNA metadata
External sram base metadata was missing from nna node. This is not the
right way to pass this data to underlying drivers. For now, a temporary
visitor is added to pass this data from the board driver. Eventually
this data should either be passed directly to the driver consuming it
from pbus or should be populated by the aml-nna driver.

Original-Bug: 342174810
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1052220
Original-Revision: 4096728a01f57034901a1ec521d285b3d20f4ba5
GitOrigin-RevId: 52bfbdf29028591f37187321cf4abd508163821a
Change-Id: Icccdeac70001c006f7df35f16f41269a68af9538