[roll] Roll fuchsia [vim3] Mali driver should take dep on Active not WakeHandling
The best practice is to take a dependency on Active when we want to
auto-suspend/resum. If we take a dependency on
ExecutionState::WakeHandling then this driver will auto-resume if any
driver wakes up and starts processing data, even if the system quickly
suspends again. This can lead to parasitic work.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1052992
Original-Revision: dcfbcca95c8b40e60a2da58ebabaf480fa30e9de
GitOrigin-RevId: ab3b3932da3d5b872224b18f37ab9192c6efad76
Change-Id: I7a1ddf9aee065d2b8535f698367f79705ca9e823