[roll] Roll fuchsia [fxfs] Improve SimpleLayerWriter code.
I discovered by chance that this code was doing deallocations as part of
writing because it was using skip() (which does a flush) to jump over a
small number of zero bytes. I've rewritten the code to fill a buffer and
write in one go. Seems a bit faster now and enabled me to write a
disk-full test to boot!. :P

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1051714
Original-Revision: b4296511c6732b7c96b44ee94fdca47d7930a6d8
GitOrigin-RevId: 7a4b6b8da20204860e9e138bace2516c8ab4d15f
Change-Id: If7f2196f1cbbd7859b2b7b6efa7eb2a829fe4c57