[roll] Roll fuchsia [netstack3] Remove uses of RngCore
The issue with using RngCore directly is that its limited API makes it
easy to accidentally commit Randomness Crimes. The primary one is
attempting to generate a uniform random value by taking a uniform u32 or
u64 and modding it to compress the range. Unless the modulus is a power
of 2, the output will no longer be uniformly distributed.

As a small example: If you have an RNG that gives you numbers uniformly
in the range [0,10], but you want numbers in the range [0,4], you end up
with the following table of mappings

RNG Out
0   0
1   1
2   2
3   3
4   0
5   1
6   2
7   3
8   0
9   1
10  2

It's clear that the range [0,2] is generated more frequently than [3,4].

I wasn't able to find any bugs while looking through the codebase,
although the ephemeral port allocation was a little dubious, since any
change to the range would cause problems. However, cleaning up this
usage means that someone copy-pasting code will be more likely to use
the Rng trait, and methods that just Do the Right Thing.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1052753
Original-Revision: 7a966e380dbd74104cbe6ed9441a04ac20f7bf50
GitOrigin-RevId: 0bfbba17975f683a03473f94db1e6af87916217c
Change-Id: I2fab281a9dd5e001608953cade70f2219d17efa7