[roll] Roll fuchsia [hrtimer][vim3] Fix lease management
jmatt@ noticed that this driver was not managing the power element
in order to get the lease satisfied. In this CL we add RequiredLevel
and CurrentLevel support to get the state of the level updated, and
we monitor the lease status before considering the lease granted.

Previous testing of lease management did not reveal this problem
because an unsatisfied lease also prevents suspension but does not
guarantees it as is the intent of this driver.

Original-Fixed: 341701622
Test: fx test -o aml-hrtimer-test
Test: hrtimer-ctl --id 0 --keep_alive 30 --wait 0 prevents suspension
      triggerd via adb shell "echo mem > /sys/power/state" for 30 seconds
      (works without this change, checking that it still works)
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1051034
Original-Revision: 2470982e2430fa4c0f45efa286a8cfe72f39981a
GitOrigin-RevId: 76aef74db88b6249955a6b73d1c0af528c6ea64d
Change-Id: I881eb3707863fcfead044f3a00079e7ca05c8e50