[roll] Roll fuchsia [kernel][gdb] Make one attempt to locate the kernel symbols.
When attempting to determine where the kernel has been relocated to
when loading symbols, make one last try to find the relocation address
using the default offset of 0x10000000 before giving up.

This script is not in great shape right now, but while it has a large
number of imperfections, it can still be very useful when attempting
to pick apart the state of a dead QEMU based system.  This hack can
keep it a bit more useful until it can (someday) be replaced with
something better.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1053753
Original-Revision: 76f742bf58fc3367da69ac8360ab42612063d357
GitOrigin-RevId: 95b488bcca4bc6024533480eba0f0d0b5d93c546
Change-Id: I79a6587605ce1c330e3ab39f54592d3f5e9f280c