[roll] Roll fuchsia [fuchsia.io] Remove Enumerate/DirectoryIterator
To speed up io2 migration, remove Enumerate (and DirectoryIterator) from
the protocol.

Enumerate/DirectoryIterator would need to be implemented as part of the
Open2 migration. There is no current server support for it and there
does not seem to be a pressing need to implement it at the moment. Most
cases are still served by ReadDirents. Removing Enumerate is to help
simplify the io2 migration process, it can be added back when migration
is completed.

Original-Bug: b/340436161
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1047692
Original-Revision: 07abe63d7f6731a6d4a8145f00c299e222e52432
GitOrigin-RevId: 1c10100ebadbbcf8c3583dd997f6035eab43bef8
Change-Id: I0be6806f62e9211170b09818e102acf5face6898