[roll] Roll fuchsia [netstack3] Avoid 'a lifetime in context_ip_bounds
If the context_ip_bounds is used to annotate an implementation that
already has a 'a lifetime in scope, the lifetimes collide and lead to
some obscure compiler errors. Avoid the confusion by giving the macro a
dedicated lifetime name that's less likely to collide with anything
declared elsewhere.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1053574
Original-Revision: 7a06246c1915361cdc225ef5e54221d553d42448
GitOrigin-RevId: 0120bff4954c388df01918c811e8b81576bd3db4
Change-Id: I73df6b735fd2594e5b8077d66b5365226ce2f5b4