[roll] Roll fuchsia [elfldltl][phdr] Refactor _dl_locked_report_globals
This is a first step in the work to disentangle libc code from
the musl dynamic linker internals.  _dl_locked_report_globals in
the musl implementation is refactored to use a new C++ subroutine
that will be reused in the integration of the new dynamic lniker.

Original-Bug: 342469121
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1054072
Original-Revision: 8d91511564fa0d30637376b7f2a4e5fae4e60190
GitOrigin-RevId: 17d4bc333e2ffd33562ae9ee24cac9fb6aa1c804
Change-Id: I3cf177a641b4ce83b391753fe5b1f25180967a13