[roll] Roll fuchsia [kernel][handle] Make observer's handle a void ptr
In the signal observer machinery the handle (class Handle) is
only used as a key. In other words the only thing that matters
is its address, to match for cancellation.

While having the type explicit is helpful as documentation
it gets in the way for evolving the handle table implementation
by tying the observer machinery with the handle table types
which is not required.

TEST=current testing suffices
BUG=fxbug.dev/42077761

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1053515
Original-Revision: 014bd73fce28f3ac841f2300859e16dd995c3216
GitOrigin-RevId: 9d9170b1294c024add957c96ce4512c2c7a07db2
Change-Id: I38fe2e2c91c61cffcabd8c8f121a6adc4ce384dd