[roll] Roll fuchsia [async-loop] Switch to zx_port_cancel_key on Fuchsia >= 20
This updates the async-loop library to use the new zx_port_cancel_key()
API when building for API level 20 or above.

This means that the async_wait structure's `object` field is no longer
used by the async loop library. Many users of this library use this
field to store data for other uses, so the field is left alone. We may
consider removing this field in the future to reduce the size of this
structure in memory.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1046693
Original-Revision: 42ce4597820967b399ecc7192fffb8ac052bf9dc
GitOrigin-RevId: 9339c218f94082d285d9946d5b7e374a3ccf37ac
Change-Id: Ie8372278a536358ab1242f555b542a871640249d