[roll] Roll fuchsia [component_manager] program input dictionary additions is not optional
Removes the `Optional` wrapping program input dictionary additions,
because an empty dictionary is functionally identical to a `None` value.
This allows us to pass the dictionary of additions directly into the
dictionary construction code, where it will retain a reference to the
additions dictionary and thus not need to upgrade a component reference
when handling requests that may involve it.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1044835
Original-Revision: 7dd0ec4452b1fef85e6b25567db8298d7772e960
GitOrigin-RevId: 0accf83d31c68f5138ad0191b8eed13e383b9eb8
Change-Id: Ia8bc9b5fb747bb99aa08c24c82fc3326c30d58a3