[roll] Roll fuchsia [display][coordinator] Move dispatcher ownership to CoordinatorDriver.
Previously, in the display coordinator driver, the Controller class
creates a fdf Dispatcher as the main event loop. This causes some
different behaviors on Controller destruction: in DFv2 drivers, the
dispatcher is destroyed by the driver framework before Stop(); while
in tests, the dispatcher is destroyed by the Controller's Stop() method.

This change fixes this issue by moving dispatcher ownership to
CoordinatorDriver, which is the class that interacts with the driver
framework. Controller now only holds an unowned reference to the
dispatcher, and it requires the dispatcher be completely stopped
when Stop() is called (which is guaranteed by DFv2 driver lifecycle).
Tests now need to manually stop the dispatcher and wait until the
dispatcher is fully stopped before calling Controller::Stop().

Test: no-op on core.nelson and core.vim3.
      "display-tool squares" worked.
Test: display-coordinator-unittests
Original-Bug: 341156315
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1055692
Original-Revision: baf233a059e025c00ac53628d66956b3b9f7b38a
GitOrigin-RevId: ba108c4791cf8f4a7c5b05278570b3c5290f8ff1
Change-Id: I0265dd91a27221c140e58479b6960e7b02641a54