[roll] Roll fuchsia [netstack3] Implement Redirect NAT
Implement the entry point for invoking NAT from IP filtering hooks,
which attempts to configure NAT for a connection if it has not yet been
configured, by running the packet through the installed NAT routines,
and then if NAT has been configured, performs NAT on the packet.

Currently the only supported type of NAT is Redirect NAT, a special case
of DNAT. DNAT is configured for a connection by rewriting the source
fields of the reply tuple of a conntrack connection, which serves as a
guide to rewrite future packets seen in both directions on the
connection.

If a destination port range is specified in a Redirect action, the
netstack attempts to find a port in the range that would not cause the
rewritten reply tuple to conflict with another tuple in the conntrack
table. If it fails to find one, the packet is dropped.

Original-Bug: 333419001
Test: netstack3-filter-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1050372
Original-Revision: c44a18a0dd63ef473923567d36496787ac7a3d33
GitOrigin-RevId: 0cebd6fc28ff2a9c358778f526562005853f07e9
Change-Id: I8345d5c66c30216cd38ff92ee2f22b3041a574fe