[roll] Roll fuchsia [hrtimer][vim3] Re-enable unit tests
The previous use of fidl::ServerBindingRef in the unit testings
allowed for potentially calling the CloseHandler method (that
captures the this pointer to access the member variable
lease_requested_) after the enclosing object went out of scope.
The use of fidl::ServerBindingRef was removed in fxrev.dev/1051034
in favor of fidl::ServerBinding which does not have this problem.

Original-Fixed: 332644994
Multiply: aml-hrtimer-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1050602
Original-Revision: e2a376434fe8cd198a56585c20e1665cf150dfba
GitOrigin-RevId: efe0c75ee827a96ed1811f78fe58750874c95b96
Change-Id: I3c35ee2cfde16497b2038391365376d322e2f19f