[roll] Roll fuchsia [fastboot]: fix clippy errors
Return errors on short writes.

Fun fact: we cannot use `write_all` for this as in TCP Fastboot mode, we
write _more_ than the passed buffer (the first 8 bytes we write is the
length of the actual message we are sending), and `write_all` will panic
as it thinkss we are trying a buffer overflow as we report back that we
wrote more than we sent.

See:

https://github.com/rust-lang/futures-rs/blob/master/futures-util/src/io/write_all.rs#L33

And

https://doc.rust-lang.org/src/core/slice/mod.rs.html#1886

Instead report back if there are short writes and error

Original-Fixed: b/42177040
Original-Bug: b/42177040
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1036216
Original-Revision: 2631911827899da1f1f7c6aab5b45af3d3eb9dbf
GitOrigin-RevId: d5479c778d8851f6bb03b0fd0ac3e991064c4b0c
Change-Id: I6668597703798f4d0dbadaf185a68f1f4a4da9ba