[roll] Roll fuchsia [aml-uart] Add lease management based on timer
Introduce a timer in the driver so that the driver can hold the wake
lease for a certain time period before dropping it. The driver resets
the timer each time when the wake vector interrupt is asserted and
drops the lease then the timer times out.

Original-Bug: b/339233136
Test: New unittests and observation on begonia(The driver doesn't drop
the lease while packets coming in, and drops the lease a while after
the last packet).

Multiply: aml-uart-dfv2-test

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1049252
Original-Revision: c627122665f096fecf25556c8d4c56c638497910
GitOrigin-RevId: ff66377098aa2e0b987ada248ca037d92636bce6
Change-Id: If41a44799d7671f5313da2e8d6cc3ce03aa12f2d