[roll] Roll fuchsia [kernel][sched] Make multiple attempts to find a CPU for a thread.
Change the semantics of the "active" bit in the mp maintained active
CPU mask.  The mask may still be observed without holding lock, but to
mutate the bit corresponding to CPU/Scheduler, the Schedueler's "queue
lock" must be held.

The implication of this is that, while _not_ holding the lock, the
values in the mask amount to a "best guess", useful in
`Scheduler::FindTargetCpu` when searching for a CPU to run a given
thread on.  That said, when it finally comes to actually queue'ing a
thread onto a target CPU, we must first hold that CPU's Scheduler's
lock, and re-verify that the CPU is actually active, potentially
backing off and re-running `FindTargetCpu` in the case that our CPU
was de-activated after we selected it, but before we actually queued
our thread on it.

Original-Fixed: 342649393
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1056093
Original-Revision: 977e309873bb06a99739f2c05ff2d5e81a1b7bc9
GitOrigin-RevId: bb73031be84b453926e7336b55e69d8ffcf713be
Change-Id: I09e22cdd30cca36a574208858f9b180ec7b11929