[roll] Roll fuchsia [bazel_sdk][tests] Make the Bazel SDK test suite less quiet.
A previous CL [1] refactored the `bazel_test.py` script to be less
chatty, but this results in really poor developer experience because:

- At build time, the command (invoked from Ninja) running the test suite
  can easily take multiple minutes, and didn't print anything unless
  it completed (with or without an error).

- Worse, because it runs in the Ninja console pool, it disables the
  pending commands table updates, resulting in what looks like a
  stuck terminal for a _very_ long time.

This restores the previous behaviour of printing Bazel test updates
to the terminal during the build. Sorry for the (lack of) noise!

[1] https://fuchsia-review.googlesource.com/c/fuchsia/+/1053433

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1056472
Original-Revision: 1f2119882b6cfba32318676d4de2d4efd8c92de9
GitOrigin-RevId: 476df3ac1c29a40fcc02abf912a6df6a853a6f87
Change-Id: Ida03427cbdac1dc6c6b48b167b2e2347ba9611de