[roll] Roll fuchsia [kernel][memalloc] Ensure that a failed insertion does not modify the pool
There was a corner case within the "insert subrange" routine that
modified internal state ahead of being sure that it had sufficient nodes
to successfully follow through. Unsuccessful operations should not
modify the pool.

This bug was surfaced in a fuzz case that unsuccessfully resized a range
(implicitly using the routine in question) and then tried to resize it
again, which hit an assertion indicating that the was no longer tracked.

Original-Fixed: 42080425
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1056584
Original-Revision: 5206c693c9976bf1a803fc7fd5a69406bd73bb32
GitOrigin-RevId: 30f5b3b992c533979467ce2a4fb6a1567ab366bc
Change-Id: I4e3db0a8bc17e40f8d2d3d3297285a4ed70084ef