[roll] Roll fuchsia [netstack3] Remove box iterator from hot path
Extract with_assigned_ipv4_addr_subnets to a test utility so we don't
rely on the box in the hot path for IPv4 address selection. It's still
used in a single test, but keeping it around for symmetry with the IPv6
version which is more widely used and has more intricate logic.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1054710
Original-Revision: 60aa3bf877bf6d2b5c647979cb7303d3be0756de
GitOrigin-RevId: f2bc57c007c065103f24a17e24e6947c01dfcdc8
Change-Id: I599d472162262ae5816b942b6c2132255b780a82