[roll] Roll infra/recipes [recipes] Add flag to git clean build step to delete nested untracked repos.
Previously, when deleted a submodule project, the build would fail on
the jiri check-clean command because the directory would fail to be
removed during the patch step, because it is non-empty. The git clean
step would fail to remove the untracked file because it is nested,
adding the additional -f flag forces it to delete the untracked file.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/1056653
Original-Revision: f6666ad5e88471a710f56d778b03a3c9965d0cd8
GitOrigin-RevId: a40dd780361ee1999344600c316eb86a629a0947
Change-Id: I5693b4ac2268e695f17e2fe502344f7f1918d764