[roll] Roll fuchsia [zircon] Introduce next gated syscall to manage power level transitions
This is prototyping a syscall using a port transport to allow
information flow from the kernel to userspace and from userspace back to
the kernel.

Additionally this introduces a new topic for set_performance_info which
communicates the current power level for a given power domain. The power
level matches that provided by the scheduler when requesting a
transition with an opaque argument.

In order to keep port packet relatively small (within the 32 bytes) the
cpu mask may not be used in this environment (64 bytes). To address this
the prior syscall that sets a power domain for a given cpu mask also
provides an ID, as a way to refer to all CPU that are members of the
domain.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1011960
Original-Revision: 571b5aa98400e122c8c3db6175ba69df0098c92b
GitOrigin-RevId: 35a82323bad43315045f1c6bc876d69daf0d4508
Change-Id: Ied3c5dc8fdaf667e9bce6b5613d541288a4e6a94