[roll] Roll fuchsia [kernel] Support IRQ configuration for UART
First we introduce a new field into the zbi_dcfg_simple_t, which
repurposes an existent reserved field as a flags field from now on.

4 bits are introduced which indicate whether is level or edge triggered.
The following two bits indicate whether its active low or high, or
whether is active low-to-high or high-tolow.

Intentionally the flag field left as zero implied no configuration is
available, and will be left to whatever the platform does by default.

In some cases like RISCV, the devicetree does not provide this
information.

As of this CL, the uart IRQ information is obtained from the devicetree,
and when available is applied before registering the irq handler.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1052916
Original-Revision: 5a2662815a44b6e982f92d7946a2b032896d07a8
GitOrigin-RevId: 2a8ef473d20e8f534f3bd20a2f56e20636846ac0
Change-Id: I03c62461a483893c4adec22af4874d82f2a05875