[roll] Roll fuchsia [boot-mmu] Build driven boot page allocation
Instead of hard coding the value, the number of allocated pages
is determined based on the boot-mmu-config, which defines a preprocessor
symbol that is the upper bound on the kernel image size.

This allows to change this estimation based on different build
information, such as whether this is an instrumented build or not.

The values uses for the image size are obtained from the addressable
sizes that the current number of preallocatd pages can address. For
instrumentation we just used the value of riscv64 which caused failures
due to not having enough pages, and boosted it by 25% (20 -> 25MB).

As a first step, physboot will panic if the kernel size is bigger than
the max kernel image size. This allows having a direct cause to what
otherwise would have been observed as a hang from zircon.

Original-Fixed: 339565993
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1045160
Original-Revision: 738f5e0a26a636a8e6d679796535791df6c8b7db
GitOrigin-RevId: d5ed1e5a3fc0b14ca3edaf422a2d274c4472ecdc
Change-Id: I623e45307a1b6c4b68df3c5cd866cbcf23aae780