[roll] Roll fuchsia [kernel][acpi] Later ACPI init and remap
Moves ACPI initialization to happen after VM initialization. ACPI was
previously initialized early to search for serial UARTs, however this
happens in physboot now.

With ACPI initialization happening later the mappings can be
dynamically created, instead of relying on the physmap. Performing this
is future work, with current logic just re-creating a 1-1 mapping in
the physmap if necessary. This is presently a no-op, but will allow for
unmapping physmap non RAM regions in the future.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1053095
Original-Revision: 5f83120fa2508e87e871cafb8da463e577e39e68
GitOrigin-RevId: 4babe7304da0b52d7b6547a4d56805bce84e2767
Change-Id: I90993b3d415c1a7462381122edc13a411b7e6898