[roll] Roll fuchsia [netstack3] Rework device sockets for split
Rework device sockets so we can split without leaking too much
information all the way to bindings.

- Fix lock ordering for socket access, it was not observing the lock
  order types.
- Rewrite tests using Rc<RefCell<_>> instead of indices.
- While I'm here, get rid of DenseMap and use the pattern of
  HashMap<Strong, Primary> used in other sockets.
- While I'm here, rename the strong socket ID so it's not ambiguous.

We're keeping the SocketId abstraction here even though that's not
applied to the other socket types as an interesting idea to pick up
again in the future if we find that it helps.

Original-Bug: 338448771
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1055513
Original-Revision: 572530a1a6cde06850ff8ed0a52ba04cf46b96b5
GitOrigin-RevId: 9593b09b28a9607b6556fb293e49f81280133347
Change-Id: I6ef80bfc3cb23a40b0eda736cac4eb896984ff6e