[roll] Roll fuchsia [vim3-devicetree] Suppress dtc warnings related to display address
There are intentionally two display nodes hdmi and dsi which both use
the same registers. This is because we do not know at build time which
of the two displays would be connected. Ideally, it would be nice to
have both the detection and driver logic in a single driver which would
avoid the duplicate register usage warning. But there is no plan to
change these drivers and hence suppressing dtc the warning.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1057899
Original-Revision: 3ecfb246a48d09d8f1ddddf5d52c78ff90cc2e5d
GitOrigin-RevId: daeaf1e97a0b831fae90805f2ef68ada037ebd48
Change-Id: I790ab9aa6aee6b1d23930220f83329adabc5f352