[roll] Roll fuchsia [kernel] Handle memory levels in MemoryWatchdog
Moves all of the memory level handling and tracking from the PmmNode to
the MemoryWatchdog. The different memory levels, except for the OOM
state, had no meaning for the PMM and were just used to generate
signals. This was complexity and logic that did not need to be in the
PMM and is only leveraged by the MemoryWatchdog.

Having the PMM notifications be more dynamic, instead of the static
reclamation levels previously, allows additional flexibility in the
future for synthesizing more complicated pressure levels based on a
mixture of PMM free and other available memory sources.

Run-All-Tests: True

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1031612
Original-Revision: bafb4efcc6d229288731d2fdedc29ce0649cbac6
GitOrigin-RevId: c5e5c6758cf7276b40e6c4b24b389d7925fd65d8
Change-Id: Ic15e8c2d510f2c649f1d5a60cff12814adbb95de