[roll] Roll fuchsia [ffx][playground] Fix incorrect addressing of newly-allocated slots
aaaaaAAAAAAAAAAA

This little transposition meant that sometimes global variables just
wouldn't get declared, or might get declared in the wrong place, or any
number of terrible things! This code only every worked I think because
the values involved tend to all be zero for tiny examples.

We probably want some newtypes here to make this harder to do in the
future, but for now we'll settle for a fix.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1057122
Original-Revision: 038b22c812d007f725000be9fe47bb1a81f8cf84
GitOrigin-RevId: 810b666f88c87a18a8f8b67b06d0fba5c447ab20
Change-Id: I5770f91f71bd24f2c801c355e3ba961e18029b6d