[roll] Roll fuchsia [fxfs] ProfileState object is recreated on the volume for each use.
The `ProfileState` object is passed into the
`FxVolume.record_or_replay_profile()` call and dropped when done. This
is primarily to delegate which `ProfileState` implementation will be
created to the `VolumesDirectory` as the `FxVolume` itself does not know
if it is a blob volume or not.

Original-Bug: b/337321099

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1049614
Original-Revision: e863cd6df8cfc765c577736d093d33b6d635bc63
GitOrigin-RevId: 5bfe274d9e44ab0eb21f59a45ddaa6838e4332f7
Change-Id: Idf7a3ec3721174b354b85646a3336111ba686c06