[roll] Roll fuchsia [fdf] Implement bootup tracker
Add a bootup tracker in the driver manager to keep
track of when all driver bind and start activity
are completed, indicating bootup is done.

This CL adds a BootupWatcher service which the tracker
implements to notify when bootup is done.

The tracker also logs an warning if there are any
hanging binding and start requests.

Original-Bug: 333714068
Tests: bootup-tracker-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1038333
Original-Revision: 1ad0035076c34bdea3ed98b2b948a9b5ed11df09
GitOrigin-RevId: b772f7f4801306d4fd12680710b1e2e6ac81b988
Change-Id: I14789bb1968b2731b0f2550b24711a3acaea1d95