[roll] Roll fuchsia [perfcompare] Fix "--expected_sample_size" check for non-perfcompare builds
Fix the recently-added "--expected_sample_size" check so that it gets
applied in the case where only one results directory is passed to
perfcompare.py, which is the common case for when results are
displayed in non-perfcompare builds.

The early return in ComparePerf() stopped this check from being
applied in that case.  I accidentally broke the check for that case by
moving it to the end of the ComparePerf() function.

Original-Bug: 343997506
Run-All-Tests: True
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1059276
Original-Revision: 9dcca711b77c289afd84226d8d80cb4ee0fc9c0a
GitOrigin-RevId: 3f8a0918e1785d1f59ebf8463b9be48d0d7f5726
Change-Id: I4383980e71522441e454008eaf6a11fd88274b6a