[roll] Roll fuchsia [sdk][build] Apply sdk_sub_build_verbose to checked subprocesses
Previously, the gn's output in sub-builds was being captured and only
shown if the sub-build failed. This meant that print statements would
not be displayed, which could be confusing.

To address this, do not capture the output when --verbose is passed to
subbuild.py as is the case when the GN arg `sdk_sub_build_verbose` is
true.

Original-Bug: 337028633
Test: Observe "IDK_SUBBUILD_api_20_x64 [0/1](1) Regenerating ninja files" when running when `fx build final_fuchsia_idk` while `sdk_sub_build_verbose=true`.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1058718
Original-Revision: d1c7b4c04d862bd50b69140c2a032176421c10e4
GitOrigin-RevId: c12abfc58f493ce890969ec4b5123c1348c61128
Change-Id: Ie78ba47eea24ace554d49258b5e563bee9c9e3ae