[roll] Roll fuchsia [superproject] Roll antlion Fix WNM config for 5 GHz hostapd AP.
When configuring WNM features (e.g. BSS Transition Management) for
hostapd AP on 5 GHz, the WNM features were not being added to the
config. This was not noticed because the firmware used for current
Fullmac devices seems to be able to reassociate through alternate logic
when BSS Transition Management is not advertised. WNM features are not
in use yet on production devices, so this did not affect testing for
production devices.

This change corrects the error, and corrects an ordering issue in the
construction of the hostapd config. I manually verified via packet
capture that the new logic advertises WNM features correctly on both
2.4 and 5 GHz hostapd APs.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/antlion/+/1059255
Original-Revision: 1dbff7bb3b9a650a6d2ddb2910499bf1a952f21f
GitOrigin-RevId: 1f4acd3b84ea34aad06f94f3c21acf62bb70825e
Roller-URL: https://ci.chromium.org/b/8746031816848233633
CQ-Do-Not-Cancel-Tryjobs: true
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1060136
Original-Revision: bb2f71290802e345b5ec79dce55393f85e4f723f
Change-Id: I275144e32704dd00b314637d8a7300e89fa6e499