[roll] Roll fuchsia [package-directory] Use VmoFile for the meta file
MetaAsFile serves the the package's hash as a file. By writing the hash
to a VMO, VmoFile can be used to the serve the hash instead. VmoFile
supports streams which are significantly faster than
fuchsia.io/File.Read for reading the file's contents.

The downside is that the meta file will now use a full page of memory
instead of 64 bytes while being served. This should not be an issue
because the meta file is typically only around for a very short period
of time.

Original-Bug: b/327633753
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1009073
Original-Revision: 9fd1997095957cbaec99a5368053e57387848aa9
GitOrigin-RevId: 3f586fa4ce1040009cf043a6c7ee61872955b5f2
Change-Id: Ic376e75686df173d01cde18433ac8b578f32c86b