[roll] Roll fuchsia [sysmem] Don't track koid when 0
When the koid is ZX_KOID_INVALID, don't track the koid for later
logging purposes since the koid is known to not be a valid koid.

The fidl call still completes with "not found"; this doesn't change
any client-visible behavior.

BUG: b/306258175
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1058714
Original-Revision: d975b42286a6c4bd2eea6c9fadadf09784359cc7
GitOrigin-RevId: b1c7f2f438342f996a6571823c486cf282bf7d98
Change-Id: I57dbda9cbcf1718c7f8f74285f51d7069029e2fb