[roll] Roll fuchsia [fxfs] Change the way we replay the journal
This CL changes how we replay to be more intuitive/simpler at the cost
of some potential performance (during replay).

Prior to this change, we would replay the mutations first and then open
the immuatble layers.  This changes things so that we open first and
then replay the mutations.  Doing it this way means that as we replay,
we don't need to keep track of deltas to apply when we open; we can just
operate the same as we do during live operation.

The downside to this is that we have to perform extra passes over the
transactions that we need to replay.  We need two more passes: one to
apply all the root parent store's mutations, and then another one to
apply the root store's mutations.  After that, we can apply all the
other mutations.

Original-Bug: b/341928701
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1058913
Original-Revision: bd96674ed8a0f8185d008679fe700cbb3afdd1d5
GitOrigin-RevId: d78854db395d35600c0ad0cd436bf035c0f6eff3
Change-Id: Ie1abe18511d302757ab52855b6cac3b9ba41396c