[roll] Roll fuchsia [fidl] Optimize the client waker
On Fuchsia, when there are multiple wakers, we can read the channel when
we are woken and then only wake the task that needs to be woken rather
than all of them.

On host, we can't do that because there are reentrancy issues, but we
can still avoid allocating a combined waker.

This is a reland. I have rewritten how the client works when there is
only a single waker to avoid the issues that caused the reverts (namely
the raciness around `into_channel`).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1058952
Original-Revision: f9b942bf4067cad24870c47466b7ab469c7de2e7
GitOrigin-RevId: 24eddbc6841232e8501de9af8526b443ae120b92
Change-Id: Idb117e88f96b8b8aa3ad455b2cbe7a70f8471625