[roll] Roll fuchsia [display][edid] Always store a copy of EDID bytes in Edid class.
Previously, for EDIDs created from external memory, the Edid class
didn't store the raw EDID bytes, but required the EDID bytes to
outlive the Edid object. This didn't work when the EDID bytes are
stored in a temporary storage (for example, in a static_vector /
std::array stored in the stack).

This change modifies the behavior of Edid::Init(bytes, len). Now
it always stores a copy of the provided bytes in its `bytes_`
vector.

Test: no behavior change.
Test: core.vim3 (on VIM3 with AML028-30MB-A1 HDMI display) booted.
Test: display-cooridinator-unittests
Original-Bug: 344042452
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1060553
Original-Revision: 2a493b5915e2e4beb3617fcf4f72bb9690551064
GitOrigin-RevId: 6fd73fa055a417d8c383512dfdfeadfc2dc41ec7
Change-Id: Iab5bc8c0af8da8041031ec04762398ee1b06347e