[roll] Roll fuchsia [display][edid] Use cpp20::span for Edid initialization.
This change replaces the Edid::Init() method with a raw pointer
and a uint16_t size variable with the Init() method using a
cpp20::span. This removes the usage of the uint16_t type as
container size, which is discouraged by the style guide.

Test: no behavior change.
Test: core.vim3 (with AML028-30MB-A1 HDMI display) booted.
Original-Bug: 344042452
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1060379
Original-Revision: d498a3d6c39306356af97b47069f6541a728bdd3
GitOrigin-RevId: 8efdc318f18db287377e69602c6beecea9f2abd9
Change-Id: I7a5ec5689b5997584a8ef68df6ef343773c844c6