[roll] Roll fuchsia [ffx][playground] Run the tests for the playground tool
The tool is experimental but we should still run the tests even though we
haven't enabled it.

We also fix a circular reference issue in cf_fs that was making the
tests fail lsan. We do that in this CL so that the change is verified by
CQ.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1058493
Original-Revision: f5f7aab99cf17f975ff2e4c7e6641f2c5a50c4a1
GitOrigin-RevId: 58ff53ac34797ef9098d952962e6964eab8d729b
Change-Id: I5d1a22e074b03f92b75f1d4e7b0f061d8199560a