[roll] Roll fuchsia [versioning] Support HEAD in ZX_*_SINCE() macros
Follow the pattern introduced in https://fxrev.dev/991253 to support
named API levels. This has the effect of limiting the possible values to
the same set of known levels.

The primary motivation for this is to make it possible to use these
macros with NEXT (https://fxbug.dev/326277078), which will be required
for making API changes once https://fxbug.dev/323889271 is implemented.

Add support for additional levels as required by existing instances of
the ZX_*_SINCE() macros. Test-specific macros for some of these levels
are no longer necessary and thus removed.

Original-Bug: 42084512
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1046234
Original-Revision: b935bbe439a0c66880584309556af2ec72d34794
GitOrigin-RevId: eaba5af6111dbbb008788d2484901998bba0042d
Change-Id: I4cb49b4853bd2a15b4a9a9e2aec17f07ddb96e3a