[roll] Roll fuchsia [netstack3] Require DeviceIdContext in filtering
Currently a lot of code in the filtering crate is parameterized on a
generic device ID type that is constrained to implement
InterfaceProperties so it can be matched against. Given that in
practice, all this code is dealing with the same device ID type, and we
sometimes want to constrain that type further (e.g. with Debug),
centralize the type and the constraints on it by requiring that
FilterIpContext, FilterHandler, and FilterHandlerProvider all implement
DeviceIdContext (this was already true for NatContext). Further, require
that the associated DeviceId type implement InterfaceProperties.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1060212
Original-Revision: bde9e382754cde30a6211ceb0b977a39a179eef1
GitOrigin-RevId: 9aec4996075d90fca9d8e82fd7fc081ece87b49a
Change-Id: I521e757ed548508afe85d8c05337dd8249802f88