[roll] Roll fuchsia [vfs][rust] Remove flags from TokenRegistry
The flags are unused.

Tokens are used to identify the destination directory of a mutating
operation. The connection that creates the token ensures that the
connection has the right to modify the directory. This avoids the
connection receiving the token from having to check the rights of the
connection that created the token.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1058592
Original-Revision: 7ce22ca0db6a692f8025cce39eb754c57c9dc2c3
GitOrigin-RevId: 607d7c2002aadeb2999df156b9cb425fb8f21cd4
Change-Id: I3b26db8f87f0622785350150aabb84090ff46cc2