[roll] Roll fuchsia [idk] subbuild.py: Add --setup-only option.
Add an option to the script to allow skipping the Ninja build
part. This will be used in the future when setting up the sub-build
repositories in a Bazel repository rule (but performing the build
separately through a Bazel action).

+ Make `--prebuilt-host-toos-dir` optional.
+ Remove an obsolete assertion.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1062134
Original-Revision: 432137375f77e2344e9ece3e586fd7f6bb7e0b5a
GitOrigin-RevId: eca8c73c973d363917b7256e10bb04bd4e6864d5
Change-Id: Ief2acb3f6ac391723849f20f71f04a53650ce1ac