[roll] Roll fuchsia [netstack3] Introduce DatagramApi
This allows us to export less symbols in the crate's public API as well
as get rid of a lot of very verbose callsites.

This is mostly all code movement so that all the functions that are pub
fit in the same block. Care was taken to try and keep them in the same
order where possible, but given the spread in the file a lot of lines
are touched.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1061639
Original-Revision: ca468e199091dac7be0d6653b65f6b273d76d91e
GitOrigin-RevId: 1371ea64aa44d4573ce3c090acfb731e303f8917
Change-Id: I62be892286c6457fef893db01bb515717edd7066