[roll] Roll fuchsia [netstack3] Remove vector allocation in addr check
The blanket impl for BaseIpTransportContext was allocating a vector to
satisfy the trait's signature, but we can do it without the allocation.

Change the shape to the `with_` form, so there's no impedance mismatch
with the traits this builds on and introduce a tailored iterator to
solve the problem for the blanket impl.

While I'm here, remove a now unnecessary lifetime bound on the devices
and addresses iterator in `IpDeviceContext`.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1061920
Original-Revision: 734520b84c942cb21eb5997236f7a7868c37c1ff
GitOrigin-RevId: 503d52faf2da4116b9ce719a2fbba125e76a8247
Change-Id: I62f32eadf33d68f13ea3bcdf2981262c86ade7b8