[roll] Roll infra/recipes [testsharder][testing_requests] Set -expects-ssh flag in botanist.
This should go after fxrev.dev/1046001 starts setting the field in the
shards. This will remove the reliance on the bringup product to
determine whether a shard should use ssh to run tests or not, and expect
the build to properly set the correct expects_ssh field in the test
spec.

Run-All-Tests: true
Original-Bug: 338300717
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/1045623
Original-Revision: 467ebc879bebf826dbb14dfc94525a7a5635f717
GitOrigin-RevId: 3b5ad4aa947f191bddaa61ced49fbe643a13b386
Change-Id: I55c7ba63701e6bf5d41c7353048603dc241d1c3e