[roll] Roll fuchsia [bazel_sdk] Fix bazel_test.py bug for --fuchsia_sdk_dir option.
Fix a subtle bug in the implementation of the `--fuchsia_sdk_dir`.
The location of the @fuchsia_sdk and @fuchsia_sdk_common
repositories was not overriden properly when running the
bazel test command, only when performing the queries that
compute the implicit dependencies for depfile generation.

This fixes the issue by correcting the way the BazelRepositoryMap
is initialized.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1062859
Original-Revision: 6a62aed23a72114ac17fe8c5af84bdc99b0df3d0
GitOrigin-RevId: eabf60c92ae5d4ab1c9f71a03e6c26ee3b02160f
Change-Id: I9192ceb43d41d5e17a035b28f0d4663253f03351