[roll] Roll fuchsia [build] Add the concept of compilation modes
This starts the replacement of the 'is_debug' GN arg and the use of
many fine-grained GN arguments with a set of 3 easy-to-specify set of
compilation modes (similar to Bazel):
- dbg
- fastbuild
- opt

The dbg and opt modes correspond to the defaults obtained today with
'is_debug = true' and 'is_debug = false'.

The new 'fastbuild' mode is a non-debug mode that uses a new
optimization level of 'minimal'.

Original-Bug: b/339031006
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1042487
Original-Revision: b49ddf642d2874ac7e6bf2766ada9b0e1aa9051d
GitOrigin-RevId: 1432c296246c782e70d13a0af12b64def415232d
Change-Id: I8f9048fd7bbd68fd7461944304d9707b9328217b