[roll] Roll fuchsia [kernel] Teach the timer layer about boot time
This is the first change that introduces the notion of boot time to the
kernel. Boot time is a new timeline that includes the amount of time the
kernel has spent in a suspended state.

Mechanically, this change introduces several timer APIs that allow
higher level code to get the boot time and update the monotonic time
offset.

Note, however, that this change does not diverge the two timelines. In
other words, the monotonic offset and the boottime offset are kept
identical, so functionally this change should be a no-op. The point of
this CL is to ensure that the APIs have the shape we want and that the
memory order on the newly mutable mono_ticks_offset is accurate.

Original-Bug: 341785588
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1061012
Original-Revision: 6416b5b6f2971b2afc0eefdbbacf49e5176f807f
GitOrigin-RevId: 06932d7d73e79da5acb8a09505ac1ae34946194c
Change-Id: Iaf45e4d6582569d6a765ad30b1ace525f89b56f1