[roll] Roll fuchsia [bazel-sdk] do not collect resources for tools
Driver tools are binaries that get added to a driver package. They are
just cc binaries and get added to the bin/ directory. They do not bring
their own resources but we were checking for resources when we were
packaging which was causing a failure.

This CL adds a tests for tools and removes the check in the packing
logic for additional tool resources.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1063274
Original-Revision: dc2c95637349259a3f464e93e43951228abfd26e
GitOrigin-RevId: b3e287e37feebf175049e6c90f62691c4984d89f
Change-Id: If9af55ecbbd849b714fba6faceb6f7069733171e