[roll] Roll fuchsia [Lacewing] Use aemu_env instead of emu_env
as emu_env points to qemu_env (when target_cpu is not x64) and host-tests aren't supported on QEMU-arm64.

This change was partially done as part of fxr/1048513 and this CL makes
this transition across all of the Lacewing self tests.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1063555
Original-Revision: 6fb4eff8027544a200cfae85c61cf7c7cf47d469
GitOrigin-RevId: c9b5f48554deb5b721740a3820195721be18bebc
Change-Id: I42296e330238c23f2da5a2d1f4edce36033d40b4