[roll] Roll fuchsia [build] Add optimization overrides to compilation_modes
This allows the overridding of options set by the compilation_mode,
starting with the 'optimize' GN arg.

This finishes the establishing of the mechanism for specifying
compilation modes, and allowing the args that are configured
by them to be overridden by the developer.

Original-Bug: b/339031006

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1043374
Original-Revision: 447463629ad97f1dcb1c0826a45a20c363f4fb95
GitOrigin-RevId: b031979caec9f77da8b593aaea98f8532bb50654
Change-Id: I8af25676ab370807cc031d60b0e4ab579825916a