[roll] Roll fuchsia [display][coordinator] Remove async-watchdog.
The async-watchdog library provides a Watchdog class that
monitors the aliveness of a given async dispatcher by
posting periodical "update" tasks on the monitored dispatcher
and periodical "check" tasks on the Watchdog-owned loop.
In the display coordinator driver, it created an async::Loop
not managed by the driver framework, which is a discouraged
(and soon-to-be disallowed) behavior.

This change deletes the async-watchdog from the display
coordinator. Future changes will add other mechanisms to
avoid deadlocks within the display coordinator.

Test: no behavior change expected.
Test: no behavior change in the internal product on nelson.
Test: no behavior change in the internal product on VIM3.
Original-Bug: 346419450
Original-Bug: 329708151
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1063543
Original-Revision: b45e833ae94e8643d5edc787db87007b0b003128
GitOrigin-RevId: 7a2875a9c7dd3f004ab3696093b42cdceee736ad
Change-Id: I50ab7629213a8f263512d3c77080fddc37e3fe21